==> code/1_TidySpermWhale.rmd <==
---
title: "Sperm Whale Tidy Data"
author: "Yvonne Barkley"
date: "2/7/2020"
output: html_document
---
This code is also saved in 'MergeCSVs.Rmd' in the CODE folder. It is the main script for dealing with the output from the MATLAB localizer.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message = FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) #allows for finding dir different than the code's dir

```


```{r}
# library(dplyr)
library(readr)
library(stringr)
library(tidyverse)
library(here) 
```


Tidy up the sperm whale data output from MATLAB using some tidyverse functions and some less elegant brute force manipulation. It works.
It mainly works in the csv_results dir and binds the localization data from the 'peak1' and 'peak2' folders (related to number of localization peaks). 


Localizations can include one or two peaks, which creates a different number of columns from the MATLAB output. 
First, deal with them separately, then combine them.
```{r message=F}
#Load data that includes only localized acoustic encounters and sighted encounters
sperm.dat <- read_csv('./data/Spermies_LOCandSIGHTnclicks.csv') #using it to get ID label and trackline lat/lon, from SpermiesBath_20200513.xlsx
sperm.datsub <- sperm.dat[,c(3,5:7, 9:14, 19, 20)] #survey, acid, sid, time, lat, lon, click info, trackline lat/lon

sperm.datsub <- unite(sperm.datsub, "ID", Survey.Number, Acoustic.ID, Sighting.ID, remove=FALSE) #makes ID column
colnames(sperm.datsub)[colnames(sperm.datsub) %in% c("Survey.Number", "Acoustic.ID", "Sighting.ID", "start.time", "Latitude", "Longitude")] <- c('survey', 'acid', 'sid', 'UTC', 'lat_trk', 'lon_trk')

#Find which ones have 1 or 2 peaks
# fpath = ("C:\\Users\\yvers\\Documents\\CHP 2\\data\\Chp2_Localizations\\csv_results")
# peaks <- list.files(path=fpath,#"C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\Simulation\\2_simdata",
#   full.names = TRUE, pattern = "*.csv") %>%
#   lapply(read_csv, col_names=F) %>% 
#   bind_rows 
# 
# #how many separate acoustic encs?
# nencs1 <- filter(peaks, peaks$X32 == '999')
# length(unique(nencs1$X31))
# 
# nencs2 <- filter(peaks, peaks$X47 == '999')
# length(unique(nencs2$X46))

```

#Localizations with 1 peak
```{r message=F}
#Now, I manually copied/pasted the separate csvs for each encounter into the peak1 or peak2 folder in 'csv_results' Chp2_Localizations.
#Some encounters have >1 tracks, that may have different number of peaks, be aware.

fpath = ("C:\\Users\\yvers\\Documents\\CHP 2\\data\\Chp2_Localizations\\csv_results\\peak1")
peak1 <- list.files(path=fpath,#"C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\Simulation\\2_simdata",
  full.names = TRUE, pattern = "*.csv") %>%
  lapply(read_csv, col_names=F) %>% 
  bind_rows 
#remove 2nd peak
peak1 = peak1[, -c(11:16,23:28,35:37)]


colnames(peak1) <- c('sig', 'dx', 'nmin', 'maxamb', 
                     'lat_bestA', 'lon_bestA', 
                     'lat_minA', 'lon_minA','lat_maxA', 'lon_maxA',
                     'unrotx_bestA', 'unroty_bestA', 'unrotx_minA', 'unroty_minA',
                     'unrotx_maxA', 'unroty_maxA',
                     'xbestA', 'xminA', 'xmaxA', 'ybestA', 'yminA', 'ymaxA', 
                     'zbestA', 'zminA', 'zmaxA', 
                     'ydiffmin', 'ydiffmax','zdiffmin', 'zdiffmax', 'survey', 'acid',
                     'sid', 'itrk')

#double-check that the length of peak1 matches length nencs
peak1 <- unite(peak1, "ID", survey, acid, sid)

#add date and time from sperm.dat with MERGE
peak1 <- merge(peak1, sperm.datsub, by.x='ID', by.y = 'ID')


##### CHECK THIS OUT, OVER HERE!!!! ####
# Since YB went back and redid the straight line encounters to get lat/lons for the second peaks, the encounters that truly only had a single peak need to have the columns with the second peak locations dealt with...THEY ARE DEALT WITH line 63

```


#Localizations with 2 peak
```{r message=F}
fpath2 = ("C:\\Users\\yvers\\Documents\\CHP 2\\data\\Chp2_Localizations\\csv_results\\peak2")
peak2 <- list.files(path=fpath2,#"C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\Simulation\\2_simdata",
  full.names = TRUE, pattern = "*.csv") %>%
  lapply(read_csv, col_names=F) %>% 
  bind_rows 
colnames(peak2) <- c('sig', 'dx', 'nmin', 'maxamb', 
                     'lat_bestA', 'lon_bestA', 'lat_bestB', 'lon_bestB', 
                     'lat_minA', 'lon_minA', 'lat_maxA', 'lon_maxA',  
                     'lat_minB', 'lon_minB', 'lat_maxB', 'lon_maxB',
                     'unrotx_bestA', 'unroty_bestA', 'unrotx_minA', 'unroty_minA', 'unrotx_maxA', 'unroty_maxA',
                     'unrotx_bestB', 'unroty_bestB', 'unrotx_minB', 'unroty_minB', 'unrotx_maxB', 'unroty_maxB',
                     'xbestA', 'xminA', 'xmaxA', 
                     'ybestA', 'yminA', 'ymaxA', 'ybestB', 'yminB', 'ymaxB',
                     'zbestA', 'zminA', 'zmaxA', 
                     'ydiffmin', 'ydiffmax', 'zdiffmin', 'zdiffmax', 'survey', 'acid', 'sid', 'itrk')

peak2 <- unite(peak2, "ID", survey, acid, sid)

#add date and time from sperm.dat with MERGE
peak2 <- merge(peak2, sperm.datsub, by.x='ID', by.y = 'ID')


```

#1. Combine Localized Encounters
```{r}

pmAll <- bind_rows(peak1, peak2, .id = 'peaks')

# pmAll2 <- unite(pmAll, "ID", survey, acid)
pmAll2 <- dplyr::select(pmAll, "ID", everything())

#consolidate data in long format
pmTidy1 <- dplyr::select(pmAll2, 'ID', 'acid', 'sid', 'UTC', lat = 'lat_bestA', lon = 'lon_bestA', itrk = 'itrk', survey = 'survey', pdist = 'ybestA')
pmTidy1$type <- 'best'
pmTidy1$peak <- 'A'

pmTidy2 <- dplyr::select(pmAll2, 'ID', 'acid', 'sid', 'UTC', lat = 'lat_minA', lon = 'lon_minA', itrk = 'itrk', survey = 'survey', pdist = 'yminA')
pmTidy2$type <- 'min'
pmTidy2$peak <- 'A'

pmTidy3 <- dplyr::select(pmAll2, 'ID', 'acid', 'sid','UTC', lat = 'lat_maxA', lon = 'lon_maxA', itrk = 'itrk', survey = 'survey', pdist = 'ymaxA')
pmTidy3$type <- 'max'
pmTidy3$peak <- 'A'

pmTidy4 <- dplyr::select(pmAll2, 'ID', 'acid', 'sid', 'UTC', lat = 'lat_bestB', lon = 'lon_bestB', itrk = 'itrk', survey = 'survey', pdist = 'ybestB')
pmTidy4$type <- 'best'
pmTidy4$peak <- 'B'

pmTidy5 <- dplyr::select(pmAll2, 'ID', 'acid','sid', 'UTC', lat = 'lat_minB', lon = 'lon_minB', itrk = 'itrk', survey = 'survey', pdist = 'yminB')
pmTidy4$type <- 'best'
pmTidy5$type <- 'min'
pmTidy5$peak <- 'B'

pmTidy6 <- dplyr::select(pmAll2, 'ID','acid', 'sid', 'ID', 'UTC', lat = 'lat_maxB', lon = 'lon_maxB', itrk = 'itrk', survey = 'survey', pdist = 'ymaxB')
pmTidy6$type <- 'max'
pmTidy6$peak <- 'B'

pmTidy <- bind_rows(pmTidy1, pmTidy2, pmTidy3, pmTidy4, pmTidy5, pmTidy6)
pmTidy <- dplyr::select(pmTidy, 'survey', 'acid', 'sid', 'ID', 'UTC', 'peak', 'itrk', 'type', 'lat' ,'lon', 'pdist')

#removes rows with NAs
# pmTidy <- pmTidy[complete.cases(pmTidy), ]


#REMOVE 1303_24, no clicks -> 4/24/20 YB removed it from folder and csv instead
# pmTidy <- filter(pmTidy, ID != '1303_24')
#ALSO NEED TO REMOVE 1705_112 peak A, type max bc it's on Maui
pmTidy <- filter(pmTidy, ID != '1705_112' | peak !='A')

#convert pdist to km
pmTidy$pdist <- round(pmTidy$pdist/1000, 2)

pmTidy <- merge(pmTidy, sperm.datsub[, c(1, 7:13)], by='ID') #specify which cols to include from sperm.datsub [, c("ID", "lat_trk", "lon_trk")]

#Adding col to indicate if localized for easier filtering
pmTidy$loc <- 1


# write.csv(pmTidy,  'C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\output/Spermies_20200424.csv', row.names = F)
# pmTidy is the input for the 2_PmEnvData code 


```

#2. Sighted encounters - extract distance and group size
```{r}
#add visual distances to pmTidyv from Spermies_SIGHTdata (subset from Amanda's file)
sw.vis <- read.csv(here::here('data', 'Spermies_SIGHTS_HI_1986-2017.csv'))

#filter for sightings from sperm.datsub
# vid <- sperm.dat[,c(3,5:7, 19, 20)]
vid2 <-filter(sperm.datsub, sperm.datsub$sid < 999)
vid3 <- unite(vid2, "ID", survey, acid, sid, remove=FALSE)
pmTidyv <- dplyr::select(vid3, ID, everything())

#NO NEED, group size in sperm.dat
#make new ID column with survey and sid for matching data purposes
id2 <- unite(pmTidyv, "ID2", survey, sid, remove=FALSE)
id2v <- unite(sw.vis, "ID2", CruzNo, SightNo)


#merge/combine distances (not group sizes) from sighting data with times and lat/lons from acoustic data
pmTidyv2 <- merge(id2, id2v[, c("ID2", "distance")], by='ID2')
pmTidyv2 <- dplyr::rename(pmTidyv2, pdist = distance)
pmTidyv2 <- pmTidyv2[,-1]
pmTidyv2$loc <- 0
pmTidyv2$peak <-'trk'
pmTidyv2$itrk <-1
```

#3. Unlocalized encounters
```{r}
sw.unloc <- read.csv(here::here('data', 'Spermies_UNLOCALIZED.csv'))
sw.unloc2 <- sw.unloc[,c(3,5:7,14,15)]
sw.unloc3 <- unite(sw.unloc2, "ID", Survey.Number, Acoustic.ID, Sighting.ID)
pmTidyun <- cbind(sw.unloc2[1:3], sw.unloc3)
colnames(pmTidyun) <- c('survey', 'acid', 'sid', 'ID', 'UTC', 'lat_trk', 'lon_trk')
pmTidyun$loc <- 0
pmTidyun$peak <-'trk'
pmTidyun$itrk <-1
```

#4. Combine all sperm whale data, localized, sighted, and unlocalized
```{r}
#combine only localized and sighted encounters
pmTidyLOCandSIGHT <- bind_rows(pmTidy, pmTidyv2)
#sort columns by survey and acID numbers
pmTidyLOCandSIGHT <- arrange(pmTidyLOCandSIGHT, survey, acid)

#combine all encounters
pmTidyALL <- bind_rows(pmTidy, pmTidyv2, pmTidyun)

pmTidyALL <- arrange(pmTidyALL, survey, acid)


pmTidyALL<-filter(pmTidyALL, ID != '1705_112_999' | peak !='A' , ID != '1706_49_999' | peak != 'A')


```

#5. Pull in aux data (nmin, nclk, click type, etc.)
#add in aux data (nmin, nclk, etc.)
```{r}
#Add nmin, nclk, and aux data
require(openxlsx)
nmin_all <- read.xlsx(paste0(here::here('data'), '/', 'SpermiesBath_20200525.xlsx'), sheet=3)
nmin <- unite(nmin_all, "ID", survey, acid, sid, remove = F)

# nclicks <- read.xlsx(paste0(here::here('data'), '/', 'SpermiesBath_20200525.xlsx'), sheet=2) #already loaded in spermdat
# nclicks <- unite(nclicks, "ID", survey, acid, sid, remove = F)


pmTidyaux1 <- merge(pmTidyALL,nmin[,c(3,7,10,13,15)], by='ID') #.x='ID', by.y = 'ID')
pmTidyaux1 <- arrange(pmTidyaux1, survey, acid)


#filter for localized and sighted/unlocalized encounters
sw.loc <- filter(pmTidyaux1, pmTidyaux1$loc==1)
sw.unlocvis <- filter(pmTidyaux1, pmTidyaux1$loc==0)
#include trackline lat/lon in the lat/lon column
sw.unlocvis$lat <- sw.unlocvis$lat_trk
sw.unlocvis$lon <- sw.unlocvis$lon_trk



#need 0-360 longitude for some env datasets. Need to make it for both localized and trackline locations
sw.loc$lon2 = ifelse(sw.loc$lon <1, sw.loc$lon + 360, sw.loc$lon) 
sw.loc$lon_trk2 = ifelse(sw.loc$lon_trk <1, sw.loc$lon_trk + 360, sw.loc$lon_trk) 

sw.unlocvis$lon2 = ifelse(sw.unlocvis$lon_trk <1, sw.unlocvis$lon_trk + 360, sw.unlocvis$lon_trk)
sw.unlocvis$lon_trk2 = ifelse(sw.unlocvis$lon_trk <1, sw.unlocvis$lon_trk + 360, sw.unlocvis$lon_trk)


#combine dfs. The 'unlocvis' where loc=0 should have the same values for lat/lon and lat_trk/lon_trk as well as the same values for lon2 and lon_trk2
pmTidyaux2 <- bind_rows(sw.loc, sw.unlocvis)# 


#drop NAs
pmTidyFinal<- pmTidyaux2 %>% drop_na(lat)

write.csv(pmTidyFinal,  'C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\output/SpermiesFinal_20200531.csv', row.names = F)
```







```
