==> code/2_NonParametrics.Rmd <==
---
title: "Wilcoxon Test between Env Data for Trackline and Localized Whales"
author: "Yvonne Barkley"
date: "5/31/2020"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---

```{r message=F}
library(lubridate)
library(tidyverse)
library(reshape2)
library(dplyr)
library(ggplot2)
library(here) #helps with stupid root dir and accessing subfolders, it's awesome.
library(rstatix)
library(broom)

```

```{r}
gridsize = 10
loctype = 'allsw'
loctype2 = 'all'
```

#Wilcoxon Test May 2020
Ran tests for prelim look at env data between localized whale locations and their trackline locations.
Didn't find much difference.
```{r}
swtestagain <- rbind(swpeakmean2, swtest_sub)
wilcoxn <- lapply(swtestagain[,c("sstAQ_m", "temp600C","chla_m","par_m","bath","slp_deg","asp_deg", "d2land_km", "d2smt_km",  "wind_ms","wavepow","ssh","sshsd","eke","group")], function(x) wilcox.test(x ~ swtestagain$group))

```

Run summary stats on variable
```{r}
swSST.long %>% 
  group_by(group) %>%
  dplyr::summarise(n=n(), 
            mean=mean(sst, na.rm = TRUE), 
            sd=sd(sst, na.rm = TRUE),
            stderr=sd/sqrt(n),
            LCL = mean - qt(1 - (0.05 / 2), n - 1) * stderr,
            UCL = mean + qt(1 - (0.05 / 2), n - 1) * stderr,
            median=median(sst, na.rm = TRUE),
            min=min(sst, na.rm = TRUE), 
            max=max(sst, na.rm = TRUE),
            IQR=IQR(sst, na.rm = TRUE),
            LCLmed = MedianCI(sst, na.rm=TRUE)[2],
            UCLmed = MedianCI(sst, na.rm=TRUE)[3])
#visualize
bxp <- ggpaired(swSST.long, x = "group", y = "sst", 
         order = c("trkSST", "peakSST"),
         ylab = "SST", xlab = "Groups")

# The test assumes that differences between paired samples should be distributed symmetrically around the median.
# Compute the differences between pairs and create histograms:

swsst2 <- swsst_sub %>% mutate(differences = peakSST - trkSST)
gghistogram(swsst2, x = "differences", y = "..density..", 
            fill = "steelblue",bins = 5, add_density = TRUE)
#If differences look symmetrical around the median, then use Wilcoxon signed-rank test

#Question : Is there any significant changes in the weights of mice after treatment?
stat.test <- swSST.long  %>%
  wilcox_test(sst ~ group, paired = TRUE) %>%
  add_significance()
stat.test

#effect size (if there is one?)
swSST.long  %>%
  rstatix::wilcox_effsize(sst ~ group, paired = TRUE)
#small effect size detected, r = 0.006

```



#Q2: Do demographic groups differ in any env data?
Use BROOM package to look at multiple env covariates in non-parametric tests between different groups defined as:

males (s) vs social groups (cd)
foraging (r, ck) vs not foraging
single animal vs multiple animals

```{r}

PmScaled <- readRDS(here::here( paste0('output/envData/complete env data/', 'CompletePm_', gridsize, 'km_', loctype2, '_scaled.rda') ))  #same as PmScaled
#Demographic Division of Data

##For singles vs groups, only use confirmed data from localized and sighted encounters, not unlocalized, bc not enough clicks or info from encounter other than a whale was present but couldn't be localized. The 'confirmed' data are either localized so we can see the TDOA tracks, or they are sighted and were counted by observers.

PmConfrmd <- filter(PmScaled, loc == 1 | loc == 0 & sid < 999 )

PmConfrmd <- PmConfrmd %>% replace_na(list(grpsize=0, nmin = 0)) #fill nas with 0 in these columns

#designate whether the encounters were a group (1) or single animal (0)
PmConfrmd$group <- 1
PmConfrmd$group[PmConfrmd$grpsize == 1 | PmConfrmd$nmin == 1] <- 0

##For foraging groups or other (slow, no foraging clicks at all), Use all whales.
#Also added in a column for the presence of a male

PmForage <- filter(PmScaled, pa >0 )

#1. Denote forage or not. 
PmFor1 <- filter(PmForage, grepl('r|k', click_code))
PmFor0 <- filter(PmForage, !grepl('r|k', click_code))
PmFor1$forage <- 1
PmFor0$forage <- 0
PmForage <- bind_rows(PmFor1, PmFor0)

#2. Denote single male or social group. Single male is defined as having one track of TDOAs and slow clicks, or a visual count of 1 animal
PmMale <- filter(PmForage, grepl('s', click_code) & nmin == 1 | grpsize == 1)
PmMale <- filter(PmMale, !grepl('cd|k|r', click_code))
PmMale0 <- filter(PmForage, !grepl('s', click_code))
PmMale$male <- 1
PmMale0$male <- 0
PmMale <- bind_rows(PmMale, PmMale0)

#3. Denote  group or not
PmForage <- PmForage %>% replace_na(list(grpsize=0, nmin = 0))
#designate whether the confirmed encounters were a group (2) or single animal (1) and 0 for unlocalized
PmForage$group <- 2
PmForage$group[PmForage$grpsize == 1 | PmForage$nmin == 1] <- 1
PmForage$group[PmForage$nmin >= 1 & PmForage$loc==0] <- 0  #can't determine group/single from unlocalized/unsighted encounter, but at least 1 whale was present

PmDemo <- PmForage

```



```
