==> code/3_Effort.Rmd <==
---
title: "Sperm Whale Habitat Model: Effort using straightPath"
author: "Yvonne Barkley"
date: "03/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---
#####
This markdown uses the function straightPath from the PAMmisc package, developed by Taiki Sakai, to calculate the amount of trackline that is straight. Certain parameters must be considered, including nSmall for determining where the ship is going, and nLarge for determining where the ship has been in order to decipher whether the ship has turned. The thld parameter indicates the amount of change occurs in the trackline to be considered different, and therefore, a turning segment. 

Yvonne has experimented with the function and found that each GPS data set can be different in several ways, including

* CHECK THE VISUAL EFFORT COLUMN, join.veffort, FOR WHITE SPACE AFTER THE STRINGS. SUCH A HEADACHE WHEN THIS HAPPENS LATER FOR FILTERING.
  unique(gpsAll$join.veffort)
  unique(gpsAll$join.aeffort)
  TRIM WHITE SPACE IF NEEDED
  gpsAll$join.veffort <- str_trim(gpsAll$join.veffort, side = "right")
  gpsAll$join.aeffort <- str_trim(gpsAll$join.aeffort, side = "right")

* UTC format: It is easy to use lubridates handy POSIXct conversion functions, BUT UTC must be converted to a character, and the correct date/time format must be realized. For example, ymd_hms vs mdy_hms.

* Are seconds included?: Older datasets don't have seconds included and must be extracted from the GPSTime column. A simple loop is included for data collected before 2016. After 2016 seems to be fine.

* Time interval: Take a look at how often the data were collected. It can range between 10 s - 1 min. 

General rules of thumb for selecting each parameter for a good starting point and running the function in general:

* nSmall: A value that equals about 2 minutes based on the time interval of data collection
* nLarge: A value that equals between 6-8 minutes
* thld:   YB kept this at 20 to be more conservative
* Run the entire gps data set THEN filter the data for lat, lon, speed, etc. This will prevent any skewed results and misinterpretations.

#Libraries
```{r}
library(PAMmisc)

library(here)
library(lubridate) #needs character or numeric to work
library(tidyverse)
library(RcppRoll)
library(data.table)

#function to split last 'n' elements in string
substrRight <- function(x, n){
  substr(x, nchar(x)-n+1, nchar(x))
}


```


#1303 - 20 second intervals
```{r}
#Open gps data for straightPath function

dir_str8path = 'C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\output\\straightPath results\\'
gps1303 <- read.csv(paste0(here('data'), '/GPSsec_1303.csv'))
```

Survey 1303 & 1641 do not have seconds included in UTC :(
    Have to manipulate the GPSTime column to extract the seconds
#1303 add seconds
```{r}
pattsec <- as.character(substrRight(gps1303$GPSTime, 2))

#some values of GPSTime appear as a single number with no '0' so it effs up the UTC. Gotta add the '0'
pattsec2 <- NULL
for (sec in pattsec){
  if (nchar(sec) == 1) {
    s = paste0('0', sec)
    pattsec2 <- rbind(pattsec2, s)
  } else
      pattsec2 <- rbind(pattsec2, sec)
  }
  
#combine UTC with seconds and convert to POSIXct
gps1303$UTC = mdy_hms(paste0(gps1303$UTC, ':', as.character(pattsec2)))
write.csv(gps1303, here::here('data', 'GPSsec_1303.csv'), row.names = F)
```

#Straight path for 1303
```{r}
nSmall = 5
nLarge = 26
thld = 20

str8_1303 <- straightPath(gps1303, nSmall, nLarge, thld, plot=TRUE)

str8_1303zoom <- filter(str8_1303, Longitude >= -162.5 & Longitude <= -161.8 & Latitude < 26)

write.csv(str8_1303, here::here('output', 'straightPath_1303.csv'), row.names = F)


#filter speeds of 8-10.5 kts from str8AllSpd
str8sub1303 <- filter(str8_1303, Speed >=7 & Speed <= 11)

myPlot <- ggplot(str8sub1303, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1303$timeDiff[1], 's'))
print(myPlot)


write.csv(str8sub1303, here::here('output', 'straightPath_1303.csv'), row.names = F)
```

#1641 - 1 min interval
```{r}
# Load data

gps1641 <- read.csv(here::here('data', 'GPS_1641.csv'))

#add 0-360 longitude
gps1641$Longitude2 <- ifelse(gps1641$Longitude <1, gps1641$Longitude + 360, gps1641$Longitude)
# subset for locations west of -150 W, into the HI EEZ
gps1641 <- filter(gps1641,  Longitude2 <= 210)

```

#1641 add seconds
```{r}
pattsec <- as.character(substrRight(gps1641$GPSTime, 2))

#some values of GPSTime appear as a single number with no '0' so it effs up the UTC. Gotta add the '0'
pattsec2 <- NULL
for (sec in pattsec){
  if (nchar(sec) == 1) {
    s = paste0('0', sec)
    pattsec2 <- rbind(pattsec2, s)
  } else
      pattsec2 <- rbind(pattsec2, sec)
  }
  
#combine UTC with seconds and convert to POSIXct
gps1641$UTC = mdy_hms(paste0(gps1641$UTC, ':', as.character(pattsec2)))

```


#Straight path for 1641
```{r}
# nSmall = 5
# nLarge = 15
thld = 20
nSmall = 2
nLarge = 7


str8_1641 <- straightPath(gps1641, nSmall, nLarge, thld, plot=TRUE)

#change lon to 0-360
str8_1641$Longitude <- ifelse(str8_1641$Longitude <1, str8_1641$Longitude + 360, str8_1641$Longitude)


#zoom in
str8_1641zoom <- filter(str8_1641, Longitude >= -165 & Longitude <= -164 & Latitude < 23.5 & Latitude >23)
gps1641zoom <- filter(gps1641, Longitude2 <= 181 & Latitude >= 30 & Latitude <= 31)
str8SWzoom <- filter(str8SW, lon2 <= 181 & lat >=30 & lat<=31)


#filter speeds of 8-10.5 kts from str8AllSpd
str8sub1641 <- filter(str8_1641, Speed >=7 & Speed <= 11)# & Longitude > -165 & Longitude < -162
                      #& Latitude < 24 & Latitude > 21)


myPlot <- ggplot(str8sub1641, aes(x=Longitude2, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
  scale_color_manual(values = c('red', 'darkgreen')) +
    ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1641$timeDiff[1], 's')) +
  geom_point(data=str8SW, mapping = aes(lon2, lat), shape = 15)
# scale_size_manual(values = c(2, 1))
print(myPlot)

#check all gps
(mygpsPLOT <- ggplot(gps1641, aes(x=Longitude2, y=Latitude)) +
  geom_point(col='black', size = 1.3) +
    ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1641$timeDiff[1], 's')) +
  geom_point(data=str8SWzoom, mapping = aes(lon2, lat), shape = 15, col='red'))

write.csv(str8sub1641, here::here('output', 'straightPath_1641.csv'), row.names = F)
```



#1604/1642 - 10 second interval
Seconds are included in UTC
```{r}
# Load data

gps1604 <- read.csv(here::here('data', 'GPS_1604.csv'))
gps1604 <- read.csv(here::here('data', 'GPS_1642.csv'))

# subset for locations west of -150 W, into the HI EEZ
gps1604$UTC <- as.character(gps1604$UTC)
gps1604$UTC <- ymd_hms(gps1604$UTC)

```

#Straight path for 1604/1642
YB added 1642 on 9/19/20
```{r}
nSmall = 15
nLarge = 35
thld = 20

str8_1604 <- straightPath(gps1604, nSmall, nLarge, thld, plot=TRUE)

#filter to zoom in on some wiggly parts
str8_1604zoom <- filter(str8_1604, Longitude >= -158.35 & Longitude <= -158.2 & Latitude < 18.8 & Latitude >18.65)


#filter speeds of 8-10.5 kts from str8AllSpd
str8sub1604 <- filter(str8_1604, Speed >=7 & Speed <= 11)# & Longitude > -165 & Longitude < -162
                      #& Latitude < 24 & Latitude > 21)


#Get shape files for Hawaii
us<-readShapePoly("C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\yb_maps\\shp\\tl_2016_us_state.shp")
hawaii<-subset(us,NAME=="Hawaii")
hawaiiMap<-fortify(hawaii) #converts shape file in a readable format for ggplot
colnames(hawaiiMap)[1] <- 'Longitude'
colnames(hawaiiMap)[2] <- 'Latitude'

hawaiiMap2 = filter(hawaiiMap, Longitude >= min(str8_1604$Longitude) & Longitude <= max(str8_1604$Longitude))

hawaiiMap2 = filter(hawaiiMap, Longitude >= min(str8_1604zoom$Longitude) & Longitude <= max(str8_1604zoom$Longitude) & Latitude >= min(str8_1604zoom$Latitude) & Latitude <= max(str8_1604zoom$Latitude))


myPlot <- ggplot(str8sub1604, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  geom_polygon(data = hawaiiMap2, aes(x = Longitude, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1604$timeDiff[1], 's'))
# scale_size_manual(values = c(2, 1))
print(myPlot)

write.csv(str8sub1604, here::here('output', 'straightPath_1604b.csv'), row.names = F)
```


#1705 - 10 second interval
Seconds are included in UTC
```{r}
# Load data

gps1705 <- read.csv(here::here('data', 'GPS_1705.csv'))

# subset for locations west of -150 W, into the HI EEZ
gps1705$UTC <- as.character(gps1705$UTC)
gps1705$UTC <- mdy_hms(gps1705$UTC)

gps1705 <- filter(gps1705, Longitude <= -150 )

```

#Straight path for 1705
```{r}
nSmall = 18
nLarge = 30#42
thld = 20

str8_1705 <- straightPath(gps1705, nSmall, nLarge, thld, plot=TRUE)

#filter to zoom in on some wiggly parts
str8_1705zoom <- filter(str8_1705, Longitude >= -158.2 & Longitude <= -158.175 & Latitude < 21.1 & Latitude >21)

#filter speeds of 8-10.5 kts from str8AllSpd
str8sub1705 <- filter(str8_1705, Speed >=7 & Speed <= 11)# & Longitude > -165 & Longitude < -162
                      #& Latitude < 24 & Latitude > 21)
str8sub1705zoom <- filter(str8sub1705, Longitude >= -158.2 & Longitude <= -158.175 & Latitude < 21.1 & Latitude >21)


#Get shape files for Hawaii
# us<-readShapePoly("C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\yb_maps\\shp\\tl_2016_us_state.shp")
# hawaii<-subset(us,NAME=="Hawaii")
# hawaiiMap<-fortify(hawaii) #converts shape file in a readable format for ggplot
# colnames(hawaiiMap)[1] <- 'Longitude'
# colnames(hawaiiMap)[2] <- 'Latitude'
# 
# hawaiiMap2 = filter(hawaiiMap, Longitude >= min(str8_1705$Longitude) & Longitude <= max(str8_1705$Longitude))

# hawaiiMap2 = filter(hawaiiMap, Longitude >= min(str8_1705zoom$Longitude) & Longitude <= max(str8_1705zoom$Longitude) & Latitude >= min(str8_1705zoom$Latitude) & Latitude <= max(str8_1705zoom$Latitude))


myPlot <- ggplot(str8sub1705zoom, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1705$timeDiff[1], 's'))
print(myPlot)

write.csv(str8sub1705, here::here('output', 'straightPath_1705.csv'), row.names = F)
```


#1706 - 10 second interval
Seconds are included in UTC
```{r}
# Load data

gps1706 <- read.csv(here::here('data', 'GPS_1706.csv'))

# subset for locations west of -150 W, into the HI EEZ
gps1706$UTC <- as.character(gps1706$UTC)
gps1706$UTC <- mdy_hms(gps1706$UTC)

#changed longitude to 0-360 for plotting purposes
gps1706$Longitude <- ifelse(gps1706$Longitude <1, gps1706$Longitude + 360, gps1706$Longitude)

```

#Straight path for 1706
```{r}
nSmall = 18
nLarge = 42#30#42
thld = 20

str8_1706 <- straightPath(gps1706, nSmall, nLarge, thld, plot=TRUE)

#filter to zoom in on some wiggly parts
str8_1706zoom <- filter(str8_1706, Longitude >= 181.5 & Longitude <= 182 & Latitude < 28.35 & Latitude >28.25)

#filter speeds of 8-10.5 kts from str8AllSpd
str8sub1706 <- filter(str8_1706, Speed >=7 & Speed <= 11)# & Longitude > -165 & Longitude < -162
                      #& Latitude < 24 & Latitude > 21)
str8sub1706zoom <- filter(str8sub1706, Longitude >= 181.5 & Longitude <= 182 & Latitude < 28.35 & Latitude >28.25)

myPlot <- ggplot(str8sub1706zoom, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8_1706$timeDiff[1], 's'))
print(myPlot)

#change longitude back to original format
str8sub1706$Longitude <- ifelse(str8sub1706$Longitude > 180, str8sub1706$Longitude - 360, str8sub1706$Longitude)

write.csv(str8sub1706, here::here('output', 'straightPath_1706.csv'), row.names = F)
```


#CHECK PM LOCATIONS
Check that the sperm whales fall along the straight segments for each survey
```{r message=FALSE}
#load sperm whale data
SwEnvData_tot <- read_csv(here::here('output', 'SpermiesWithEnvData_20200326.csv'))

SwEnvData_tot <- read.csv('./output/SpermiesFinalLR_20200919.csv', stringsAsFactors = FALSE)
surveynum = 1303
str8segs <- read.csv(here::here('output/straightPath results', paste0('straightPathWeffort_', surveynum,'.csv')))


#1303
sw1303 <- filter(SwEnvData_tot, survey == surveynum) 

#1641
sw1641 <- filter(SwEnvData_tot, survey == surveynum) 

#1642
sw1642 <- filter(SwEnvData_tot, survey == surveynum)

#1604
sw1604 <- filter(SwEnvData_tot, survey == 1604)# & lon >= -159 & lon <= -158 & lat < 22.5 & lat > 22) 
str8sub1604zoom <- filter(str8_1604, Longitude >= -159 & Longitude <= -158 & Latitude < 22.5 & Latitude > 22)

#1705
sw1705 <- filter(SwEnvData_tot, survey == 1705) 

#1706
#change 1706 longitude to 0-360, zoom in on wherever
str8sub1706$Longitude <- ifelse(str8sub1706$Longitude <1, str8sub1706$Longitude + 360, str8sub1706$Longitude)
str8sub1706zoom <- filter(str8sub1706, Longitude >= 203 & Longitude <= 203.25 & Latitude < 20.5 & Latitude > 20.25)
sw1706 <- filter(SwEnvData_tot, survey == 1706)# & lon2 >= 203 & lon2 <= 203.25 & lat < 20.5 & lat > 20.25)


#swap the straight segments for each survey to check if the sperm whales are located along straight trackline segments

str8SW = sw1303


(myPlot <- ggplot(str8segs, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=straight), size = 1) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  
  geom_point(data=str8SW, mapping = aes(lon, lat), shape = 15) +
  ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge, " , Thld=", thld, 
                 ", TimeDiff=", str8segs$timeDiff[1], 's')))


```

#Effort Match
Determine when and where the acoustics team was ON effort along straight segments.
Use the 'str8sub' data for each survey because those include only the straight 
```{r}
survey = 1604
str8segs = eval(as.name(paste0('str8sub', survey)))
str8SW = eval(as.name(paste0('sw', survey)))

str8segs = read.csv(here::here('output/straightPath results/straightPath_1604.csv'))
eff <- read.csv(here::here('data', paste0('Effort_', survey, '.csv')))
colnames(eff) <- c("UTC", "PCTime", "cruise_number", "ac_observer", "ac_effort", "vis_effort")
# eff <- filter(eff, ot_effort == 'S')
eff$UTC <- mdy_hm(as.character(eff$UTC))
eff$PCTime <- mdy_hm(as.character(eff$PCTime))

#to check duplicates, etc.
# dup <- as.data.frame(eff$PCTime %in% str8segs$UTC) 
# eff[duplicated(eff),]
# eff[duplicated(eff$PCTime),]
# 
# 
# dup <- str8segs[duplicated(str8segs$UTC),]
# dup2 <- as.data.frame(eff$PCTime %in% dup$UTC) 


###zoom in if necessary
# str8segs <- filter(str8segs, Longitude <= 180 & Longitude <= 205 & Latitude < 22 & Latitude >21)
# str8SW <- filter(str8SW, lon2 >= 202 & lon2 <= 205 & lat < 22 & lat >21)
require(data.table)
#use data.table functions to match the times and fill in the effort to the gps dataframe
df_gps<- data.table(as.POSIXct(str8segs$UTC ))   #gps dataframe after using straightPath, filtered for speed
# df_gps=df_gps[16000:21359] #sometimes joins are uneven
colnames(df_gps) <- "date.gps"

df_eff <- data.table(as.POSIXct(eff$PCTime), eff$ac_effort, eff$vis_effort)   #effort dataframe
# df_eff <- df_eff[14:20,] #?
colnames(df_eff) <- c("date.eff", "aeffort", "veffort")

# setDT(df_gps)
# setDT(df_eff)

df_gps[, join.time := date.gps] #creates duplicate of UTC in column named 'date'
df_eff[, join.time := date.eff]

setkey(df_gps, join.time) #set column to perform the join on
setkey(df_eff, join.time)


join= df_eff[df_gps, roll = Inf]
join2= df_gps[df_eff, roll = Inf]

join[duplicated(join$date.eff)]

# join2 = df_eff[df_gps, roll = 'nearest'] #do not use


# join=join[-(1:4),] #did this for 1642 due to mismatch in row number, hopefully correct

# join[duplicated(join),]
# dup3 <- as.data.frame(table(join$UTC))
# dup3[dup3$Freq >1,]


str8eff <- data.frame(str8segs, join$aeffort, join$veffort) %>% drop_na(join.aeffort)
str8effsub <- str8eff[, c(1:3,24,25)]
#both ac and vis on
# what3 <- filter(str8eff, straight == 'TRUE' & join.veffort == 'on')# & join.veffort == 'on' | join.veffort == 'off')


#Plot results of effort with sperm whales

(myPlot2 <- ggplot(str8eff, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=join.aeffort), size = 1) +
  scale_color_manual(values = c('red', 'darkgreen')) +
  
  geom_point(data=str8SW, mapping = aes(lon2, lat), shape = 15))


write.csv(str8eff, here::here('output', paste0('straightPathWeffort_', survey, '2.csv')), row.names = FALSE)
```


#Clean up for Taiki
```{r}
#reload str8eff data sheets to remove join.veffort and rename join.aeffort
survey = 1706
str8segs = read.csv(here::here('output', paste0('straightPathWeffort_', survey, '.csv')))


str8segs = str8segs[,-c(24)]

str8segs = str8segs[,-c(12,25)]

colnames(str8segs)[23] <- 'aeffort'

write.csv(str8segs, here::here('output', paste0('straightPathWeffort_', survey, '.csv')), row.names = FALSE)


ggplot(str8segs, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=aeffort), size = 1) +
  scale_color_manual(values = c('red', 'darkgreen'))
```






#WTF
If the effort looks messed up, investigate that shit here.
```{r}
#pick out section of 'str8eff' where it may be wrong. Then check original effort data for the times.

str8effzoom <- filter(str8eff, Longitude <= 190 & Longitude >= 185 & Latitude <= 29 & Latitude >= 27.5)

ylat = c(28.62005833, 28.07774167	)
xlon = c(-173.91778, -171.6335783)
xlon = xlon+360
xychk = data.frame(xlon, ylat)

(myPlotchk <- ggplot(str8effzoom, aes(x=Longitude, y=Latitude)) +
  geom_point(aes(group=timeGroup, col=join.aeffort), size = 1) +
  scale_color_manual(values = c('red', 'darkgreen'))) +
  
  geom_point(data = xychk, mapping = aes(xlon, ylat), shape = 19)







```

#filter gps data for speeds between 8-10 knots
gps1303sub1 <- filter(gps1303, Speed >= 8 & Speed <= 10.5)
gps1303sub1$UTC2<- as.character(gps1303sub1$UTC)

#function to split last 'n' elements in string
substrRight <- function(x, n){
  substr(x, nchar(x)-n+1, nchar(x))
}
pattsec <- substrRight(gps1303sub1$GPSTime, 2)

gps1303sub1$UTC = mdy_hms(paste0(gps1303sub1$UTC, ':', pattsec))


#run straightPath from PAMmisc
require(RcppRoll)
str8path <- straightPath(gps1303sub1, 30, 90, 20, plot=TRUE)


#Explore a small subset of GPS data
gps1303sub2 <- filter(gps1303sub1, Longitude <= -175.5 & Latitude > 27.2)
str8path1 <- straightPath(gps1303sub2, 30, 60, 20, plot=TRUE)
str8path2 <- straightPath(gps1303sub2, 15, 30, 20, plot=TRUE)

write.csv(str8path2, here('output', 'TestPath_1303subset.csv'), row.names = F)
write.csv(gps1303sub2, here('output', 'TestPath_1303data.csv'), row.names = F)


big60 <- filter(str8path2, Longitude > -175.67)
big90 <- filter(str8path2, Longitude > -175.67)
big90b <- filter(str8path2, Longitude > -175.67)


#subset for longitudes west of 172
gps1303W172 <- filter(gps1303sub1, Longitude <= -172)
str8pathW172 <- straightPath(gps1303W172, 30, 60, 20, plot=TRUE)


#use all speeds within the same area
gps1303sub3 <- filter(gps1303, Longitude <= -175.5 & Latitude > 27.2)
str8pathALL <- straightPath(gps1303sub3, 30, 90, 20, plot=TRUE)


myPlot2 <- ggplot(AllSpeeds, aes(Longitude, Latitude)) +
  geom_point()
print(myPlot2)

myPlot3 <- ggplot(gps1303, aes(Longitude, Latitude)) +
  geom_path()
print(myPlot3)

patt2 <- substr(gps1303sub1$UTC2, 1, 9) #subset the longest date string
patt2trim <- str_trim(patt2, side = "right") #trim the white space off the shorter strings




dv = patt2trim

hh = floor(gps1303sub1$GPSTime/10000);
mm = floor((gps1303sub1$GPSTime-hh*10000)/100);
ss = floor((gps1303sub1$GPSTime-hh*10000-mm*100));
dv2 = paste0(dv, hh, mm, ss)


require(M3)
patt3 <- combine.date.and.time(date = dv, time=list(hh, mm, ss))
