==> code/4_PmSSHEKE.rmd <==
---
title: "Derive SSH/SSHsd/EKE from netcdf"
author: "Yvonne Barkley"
date: "4/7/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
This code uses the data from 'global-reanalysis-phy-001-030-monthly_SSHAUV.nc' in the data folder to calculate the SSHsd and EKE from u, v, ssh data. 
Data were pulled from June 2010-Dec 2017, with 92 layers in the nc file that represent each month. The date format is 'hours since Jan-1-1950'. They are converted below into 'ssh_dates'.


#Standard deviation SSH

This uses the data from CMEMS instead of GODAS. There are several papers that use this data set to get ssh, sshsd, and EKE.


```{r}
library(tidyverse)
library(tidync)
library(ncdf4)
library(raster)


flist <- list.files(path = 'C:\\Users\\yvers\\Documents\\CHP 3/SpermWhales/data', pattern = "^.*\\.(nc|NC|Nc|Nc)$")

#Use 'global-reanalysis-phy-001-030-monthly_SSHAUV.nc' from CMEMS
sshnc = paste0(here::here('data'), '/', flist[11]) #flist don't work no more
#2022: was trying this stuff for pampal code, but not necessary
sshnc = list.files(path = 'E:\\BACKUP-PhD\\Chapter 3-Sperm SDM\\code_data\\SpermWhales\\data',
                   pattern = 'nc') #flist don't work no more

sshnc = nc_open(sshnc[11])
attributes(sshnc$var)$names

nc <- nc_open(sshnc)
attributes(nc$var)$names

#check out what's in it
tidync(sshnc)

#Stack/brick (bricks linked to a single multilayer file) for each variable, uo, vo, zos
# Bricks are more efficient, read to single file vs multiple files 
#eastward velocity = uo
sshb <-brick(sshnc, varname="zos") 
# values(sshb)
# crs(sshb)

## Names of layers, which are the dates
# getting dates from bands?
# units: hours since 1950-01-01
name <- names(sshb) #name of each layer (which is the month)
namesplit <- strsplit(name, "X", fixed=T) #split off the 'X' in front of the number
namedf <- data.frame(matrix(unlist(namesplit), nrow=length(namesplit), byrow=TRUE)) #make a dataframe with the dates as names
colnames(namedf) <- c('X', 'date')

#NOTE about attempted changes
#####
# YB attempted to change things 9/9/20, but no need. Also not sure that the original code for 'namedate' works... 
# require(stringr)
# #trying to convert weird dates to actual posixct objects like "2013-01-01 08:00"
# nameperiods <- str_replace_all(namedf$date, "[.]", "-")
# 
# nameperiods %>% .[10] %>% str_pad(1, "right") %>% str_c(nameperiods[1:10])  ##SO CLOSE! NEed to add in : for HMS
# letters %>%
#   .[1:10] %>% 
#   str_pad(3, "right") %>%
#   str_c(letters[2:11])
# 
# names2 <- substring(nameperiods, 11, 11) <- ' ' #pull out CTD number in file name
#####
namedate <- as.numeric(levels(namedf$date))[namedf$date]
namedate <- as.POSIXct(namedate*3600, origin="1950-01-01",tz="GMT", format="%Y-%m-%d") #do i need this???

nout = nlayers(sshb) #layers represent months in the brick


################ SSH ####################
lon=sw$lon2


#Get the SSH
sshtotal = NULL
for (band in 1:nout) {
  start = sshb[[band]] #select a band/month
  df = as.data.frame(rasterToPoints(start)) #convert rasterlayer into df
  colnames(df) <- c("lon", "lat", band) #add col names, using the band number representing each month
  sshtotal = cbind(sshtotal, df[, 3]) #each column is the SSH for each month
  
} 
#save sshtotal
saveRDS(sshtotal, file = paste0(here::here('data'), '/', 'ssh_total.rda'))

# sshtotal <- as_tibble(cbind(df[,1:2], sshtotal)) 
#need column names to be the actual dates
sshtotaldat <- sshtotal[3:94]


colnames(sshtotaldat) <- namedate
sshtotal <- cbind(sshtotal[,1:2], sshtotaldat) #add lon lat back in
sshtidya <- sshtotal %>% pivot_longer(-(lon:lat),names_to = 'date', values_to = 'ssh')
sshtidya$date <- as.POSIXct(sshtidya$date)

#save sshtotal
saveRDS(sshtidya, file = paste0(here::here('data'), '/', 'ssh_tidy.rda'))

ssh_match = list()
for (i in 1:nrow(sw)){
 
  lontmp <- which(abs(sshtidya$lon-lon[i]) == min(abs(sshtidya$lon-lon[i])) & 
                    abs(sshtidya$lat-lat[i]) == min(abs(sshtidya$lat-lat[i])) &
                    abs(sshtidya$date-dates[i]) == min(abs(sshtidya$date-dates[i])))
  sshtmp <- sshtidya[lontmp[1],] #some had duplicates, so take the first one
  ssh_match <- rbind(ssh_match, sshtmp)
}





################ SSH SD  ####################
#Get the standard deviation of SSH
sshsdtotal5 = NULL
for (band in 1:nout) {
  start = sshb[[band]] #select a band/month
  foc = focal(start, w=matrix(1/25,5,5), fun=sd) #use foc to take stand dev of each point using 3x3 cells
  df = as.data.frame(rasterToPoints(foc)) #convert rasterlayer into df
  colnames(df) <- c("lon", "lat", band) #add col names, using the band number representing each month
  sshsdtotal5 = cbind(sshsdtotal5, df[, 3]) #each column is the SSHsd for each month
  
} 
saveRDS(sshsdtotal5, file = paste0(here::here('data'), '/', 'sshsd_total5.rda'))

sshZtotal <- as_tibble(cbind(df[,1:2], sshsdtotal)) 
#need column names to be the actual dates
sshZtotaldat <- sshZtotal[3:94]

colnames(sshZtotaldat) <- namedate
sshZtotal <- cbind(sshZtotal[,1:2], sshZtotaldat)
sshZtidya <- sshZtotal %>% pivot_longer(-(lon:lat),names_to = 'date', values_to = 'sshsd')
sshZtidya$date <- as.POSIXct(sshZtidya$date)

#save tidy version sshsdtotal
saveRDS(sshZtidya, file = paste0(here::here('data'), '/', 'sshsd_tidy.rda'))


sshsd_match = list()
for (i in 1:nrow(sw)){
 
  lontmp <- which(abs(sshZtidya$lon-lon[i]) == min(abs(sshZtidya$lon-lon[i])) & 
                    abs(sshZtidya$lat-lat[i]) == min(abs(sshZtidya$lat-lat[i])) &
                    abs(sshZtidya$date-dates[i]) == min(abs(sshZtidya$date-dates[i])))
  sshsdtmp <- sshZtidya[lontmp[1],] #some had duplicates, so take the first one
  sshsd_match <- rbind(sshsd_match, sshsdtmp)
}






```


#Plotting SSH metrics with Spermie locations
```{r}
#Only works with single raster layers, not stacks or bricks. standard deviation using 3x3 pixels
sshr <- raster(sshnc, varname = 'zos')
sshr <- sshb[[89]] # which month? Oct 2017

# sshsd <- calc(sshr, sd) #doesn't work

sshm <- calc(sshr, mean)
sshm2 <- cellStats(sshr, mean) #output is single number, not used

sshDF <- rasterToPoints((sshr)) #df of ssh for each location in raster
sshDF2 <- rasterToPoints((sshm)) #same as sshDF

#uses focal on single raster layer fo Oct 2017
sshfocsum <- focal(sshr, w=matrix(1,3,3), fun=sd)
sshfocsumDF <- rasterToPoints(sshfocsum)
as.POSIXct(594252*3600, origin="1950-01-01",tz="GMT", format="%Y-%m-%d") #check date
#Import all rasters by band #, run focal on them for sd, then save to spdf

swOct17 <- read.csv("C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\data/Spermies_Oct2017.csv", header=TRUE)
swOct17$xlon2 <- swOct17$xlon+360
dirfig = "C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\figures/"

#automatically saves figs in dirfig
png(paste0(dirfig, "SSH_Oct2017_sum.png"), width = 1000, height = 600)
plot(sshr, main = "Oct 2017 SSH Monthly Mean (m)")
par(new=TRUE)
plot(swOct17$xlon2, swOct17$ylat, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)
dev.off()

png(paste0(dirfig, "SSHsd_Oct2017_sum.png"), width = 1000, height = 600)
plot(sshfocsum, main = "Oct 2017 Standard Deviation SSH (m)")
par(new=TRUE)
plot(diffsshsddf$lon2, diffsshsddf$lat, col="black", pch=16, cex=1, xlab='', ylab='', axes=F) #big diff in sshsd

plot(swOct17$xlon2, swOct17$ylat, col="black", pch=16, cex=1,xlab='', ylab='', axes=F)

dev.off()

########################
#Aug 2010
sshr0810 <- sshb[[3]]
sshfocsum0810 <- focal(sshr0810, w=matrix(1,3,3), fun=sd)
sshfocpts0810 <- rasterToPoints(sshfocsum0810)
sshfocsumDF0810 <- as.data.frame(sshfocpts0810)
as.POSIXct(531420  *3600, origin="1950-01-01",tz="GMT", format="%Y-%m-%d") 


#plot distribution of raster

sshhist <- hist(sshr0810, breaks = 5, main = 'Histogram of SSH', col = 'purple', xlab="SSH (cm)")
sshhistsd <- hist(sshfocsum0810, breaks = 6, main = 'Histogram of SSHsd', col = 'pink', xlab="SSHsd (cm)")

int <- (sshfocsum0810@data@max-sshfocsum0810@data@min)/6

pal = colorRampPalette(c("light green", "green", 'lightgoldenrod1', "red", 'red4'))
plot(sshfocsum0810, #breaks = seq(sshfocsum0810@data@min, sshfocsum0810@data@max, by=int),
col = pal(5),
main = 'SSHsd Aug 2010')
par(new=TRUE)
plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)



# plot(sshfocsum0810, breaks = c(0.000, 0.005, 0.010, 0.015, 0.020, 0.025, 0.030), 
#      col = terrain.colors(6),
#      main = 'SSHsd')
# par(new=TRUE)
# plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)

swAug10 <- read.csv("C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\data/Spermies_Aug2010.csv", header=TRUE)
swAug10$xlon2 <- swAug10$lon_trk+360
dirfig = "C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\figures/"


png(paste0(dirfig, "SSH_Aug2010_sum.png"), width = 1000, height = 600)
plot(sshr0810, main = "Aug 2010 SSH Monthly Mean (m)")
par(new=TRUE)
plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)
dev.off()

png(paste0(dirfig, "SSHsd_Aug2010_sum.png"), width = 1000, height = 600)
plot(sshfocsum0810, main = "Aug 2010 Standard Deviation SSH (m)")
par(new=TRUE)
plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1,xlab='', ylab='', axes=F)

dev.off()


cuts= c(0.000, 0.005, 0.010, 0.015, 0.020, 0.025, 0.030)
pal = colorRampPalette(c("light green", "green", "red"))

plot(sshfocsum0810, breaks=cuts, col=pal(6))
par(new=TRUE)
plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)


```



#EKE using U & V data
U is eastward, V is northward velocities
```{r}
################ U data ####################
# (E-W, meridian currents, eastward velocities)

ub <-brick(sshnc, varname="uo") 
# values(ub)
# crs(sshb)
nout = nlayers(ub) 

utotal = NULL
for (band in 1:nout) {
  start = ub[[band]] #select a band/month
  # foc = focal(start, w=matrix(1,3,3), fun=sd) #use foc to take stand dev of each point using 3x3 cells
  df = as.data.frame(rasterToPoints(foc)) #convert rasterlayer into df
  colnames(df) <- c("lon", "lat", band) #add col names, using the band number representing each month
  utotal = cbind(utotal, df[, 3]) #each column is the SSHsd for each month
  
} 

uZtotal <- as_tibble(cbind(df[,1:2], utotal)) 
#need column names to be the actual dates
uZtotaldat <- uZtotal[3:94]
colnames(uZtotaldat) <- namedate
uZtotal <- cbind(uZtotal[,1:2], uZtotaldat)

uZtidya <- uZtotal %>% pivot_longer(-(lon:lat),names_to = 'date', values_to = 'u')



################ V data ####################
# (N-S, meridian currents, northward velocities)
vb <-brick(sshnc, varname="vo") 

vtotal = NULL
for (band in 1:nout) {
  start = vb[[band]] #select a band/month
  # foc = focal(start, w=matrix(1,3,3), fun=sd) #use foc to take stand dev of each point using 3x3 cells
  df = as.data.frame(rasterToPoints(foc)) #convert rasterlayer into df
  colnames(df) <- c("lon", "lat", band) #add col names, using the band number representing each month
  vtotal = cbind(vtotal, df[, 3]) #each column is the SSHsd for each month
  
} 
vZtotal <- as_tibble(cbind(df[,1:2], vtotal)) 
#need column names to be the actual dates
vZtotaldat <- vZtotal[3:94]
colnames(vZtotaldat) <- namedate
vZtotal <- cbind(vZtotal[,1:2], vZtotaldat)

vZtidya <- vZtotal %>% pivot_longer(-(lon:lat),names_to = 'date', values_to = 'v')


saveRDS(utotal, file = paste0(here::here('data'), '/', 'u_total.rda'))
saveRDS(vtotal, file = paste0(here::here('data'), '/', 'v_total.rda'))


################ EKE data ####################
# EKE=1/2x(U²+V²) -  u=zonal geostrophic currents, V=meridian geostrophic currents

eke = 0.5*(uZtotaldat^2 + vZtotaldat^2) #all eke values, no lon/lat

ekeZtotal <- as.data.frame(cbind(vZtotal[,1:2], eke)) 

ekeZtidya <- ekeZtotal %>% pivot_longer(-(lon:lat),names_to = 'date', values_to = 'eke')

ekeZtidya$date <- as.POSIXct(ekeZtidya$date)
saveRDS(ekeZtidya, file = paste0(here::here('data'), '/', 'eke_tidy.rda'))



eke_match = list()
for (i in 1:nrow(sw)){
 
  lontmp <- which(abs(ekeZtidya$lon-lon[i]) == min(abs(ekeZtidya$lon-lon[i])) & 
                    abs(ekeZtidya$lat-lat[i]) == min(abs(ekeZtidya$lat-lat[i])) &
                    abs(ekeZtidya$date-dates[i]) == min(abs(ekeZtidya$date-dates[i])))
  eketmp <- ekeZtidya[lontmp[1],] #some had duplicates, so take the first one
  eke_match <- rbind(eke_match, eketmp)
}

hist(eke_match$eke)


#turn df into raster. Each layer for each column of eke
eker <- rasterFromXYZ(ekeZtotal)
nlayers(eker)
 








#maps
#Aug 2010
eke0810 <- eker[[3]]
ufocsum0810 <- focal(u0810, w=matrix(1,3,3), fun=sd)

ufocpts0810 <- rasterToPoints(ufocsum0810)
ufocsumDF0810 <- as.data.frame(ufocpts0810)
as.POSIXct(531420  *3600, origin="1950-01-01",tz="GMT", format="%Y-%m-%d") 



#plot distribution of raster

uhist <- hist(ub, breaks = 5, main = 'Histogram of SSH', col = 'purple', xlab="SSH (cm)")
uhistsd <- hist(sshfocsum0810, breaks = 6, main = 'Histogram of SSHsd', col = 'pink', xlab="SSHsd (cm)")

int <- (ufocsum0810@data@max-ufocsum0810@data@min)/6


pal = colorRampPalette(c("light green", "green", 'lightgoldenrod1', "red", 'red4'))
plot(ufocsum0810, #breaks = seq(sshfocsum0810@data@min, sshfocsum0810@data@max, by=int),
col = pal(5),
main = 'U Aug 2010')
par(new=TRUE)
plot(swAug10$xlon2, swAug10$lat_trk, col="black", pch=16, cex=1, xlab='', ylab='', axes=F)

```


```{r}
require(rgdal)
require(sf)

sea <- readShapePoly('C:\\Users\\yvers\\Documents\\CHP 3/SpermWhales/data/Seafloor Features/Seamounts.shp')


sea <- spTransform(readOGR('C:\\Users\\yvers\\Documents\\CHP 3/SpermWhales/data/Seafloor Features/Seamounts.shp'), CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"))
names(sea)
plot(sea)





```




















#SST standard deviation
```{r}
sstnc = paste0(here::here('data'), '/', flist[13])
tidync(sstnc)
sstr <- raster(sstnc, varname = 'sst')
sstfoc <- focal(sstr, w=matrix(1,3,3), fun=sd)

png(paste0(dirfig, "SST_Oct2017.png"), width = 600, height = 400)
plot(sstr, main="Oct 2017 Monthly Mean SST")
par(new=TRUE)
plot(swOct17$xlon, swOct17$ylat, col="black", pch=16, cex=1,xlab='', ylab='', axes=F)
dev.off()

png(paste0(dirfig, "SSTsd_Oct2017.png"), width = 600, height = 400)
plot(sstfoc, main = "Oct 2017 Standard Deviation SST (m)")
par(new=TRUE)
plot(swOct17$xlon, swOct17$ylat, col="black", pch=19, cex=1,xlab='', ylab='', axes=F)

dev.off()







```
