==> code/5_ExploreDataPlots.rmd <==
---
title: "ExploreDataPlots"
author: "Yvonne Barkley"
date: "3/24/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE, warning=FALSE}
library(tidyverse)
library(reshape2)
library(geoR)
library(fields)
library(sp)
library(maptools)
library(rgdal)
library(classInt)
library(lattice)
library(zoo)
library(ggplot2)
library(Rmisc)
library(MuMIn)
library(matrixStats)
```


#Load data

```{r}
dir = 'C:\\Users\\yvers\\Documents\\CHP 3\\SpermWhales\\figures/'
SwEnvData_map <- read.csv(here::here('output', 'SpermiesWithEnvData_20200531.csv'))
SwEnvData_map <- read.csv(here::here('output', 'SpermiesWithEnvData_20200527-scaled.csv'))

#load data from Combined 'loctype' at 25km grid
SwEnvData_mapb <- readRDS(here::here( paste0('output/models/', loctype, '/data/', 'CompletePm_', gridsize, 'km_', loctype2, '_EDIT.rda') ))

#The simple maps hate when data crosses the dateline, so for now, I'm adjusting everything back to the eastern side of the dateline for visual purposes
SwEnvData_map$lon <- ifelse(SwEnvData_map$lon > 1, SwEnvData_map$lon*-1, SwEnvData_map$lon)

#too many click_codes, reduce creaks/regular (ck/r) to 'f' for foraging -> s, f, cd, cds, cdf, fs, cdfs
SwEnvData_map$click_code2 <- str_replace_all(SwEnvData_map$click_code, "r|ck", "f")
SwEnvData_map$click_code2 <- str_replace_all(SwEnvData_map$click_code2, "ff", "f")

#https://stackoverflow.com/questions/19747384/create-new-column-in-dataframe-based-on-partial-string-matching-other-column
SwEnvData_map$click_grp <- ifelse(grepl("c", SwEnvData_map$click_code2, ignore.case = T), "social", 
                           ifelse(grepl("s", SwEnvData_map$click_code2, ignore.case = T), "male", "forage"))
fs <- filter(SwEnvData_map, click_code2 == 'fs') #subset out then recombine
fs2 <- filter(SwEnvData_map, click_code2 != 'fs') #subset out then recombine

fs$click_grp <- str_replace(fs$click_code2, "fs", "forage/male")
SwEnvData_map <- as.data.frame(rbind(fs, fs2))

#now filter for just the 'best' location of localized and the track locations for vis and unloc
SwEnvData_mapb1 <- filter(SwEnvData_map, loc == 1, type == 'best', peak == 'A')
SwEnvData_mapb2 <- filter(SwEnvData_map, loc == 0, sid == 999)
SwEnvData_mapb3 <- filter(SwEnvData_map, loc ==0, sid < 999)
SwEnvData_mapb1$enctype <- 'loc'
SwEnvData_mapb2$enctype <- 'un'
SwEnvData_mapb3$enctype <- 'vis'

SwEnvData_mapb = as.data.frame(rbind(SwEnvData_mapb1, SwEnvData_mapb2,SwEnvData_mapb3))


```


#Get Hawaii map data and whale data
```{r}
#OLD
#Get shape files for Hawaii
# us<-readOGR("C:\\Users\\yvers\\OneDrive\\PHD\\CHP2&3-Sperm\\code\\yb_maps\\shp\\tl_2016_us_state.shp")
# hawaii<-subset(us,NAME=="Hawaii")
# hawaiiMap<-fortify(hawaii) #converts shape file in a readable format for ggplot
```
```{r}
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=as.factor(year), shape=enctype), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2)) +
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(8, 17, 15),name='Enc Type', labels=c("Ac_Localized", "Ac_Unlocalized","Sighting")) +
 scale_color_manual(values = c("#d64161", "#ff7b25",  "#86af49", "#6b5b95"), name='Year') + #feb236", "#92a8d1",
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("All Sperm Whale Encounters") +
  theme(legend.position="bottom", legend.box='horizontal',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))


ggsave(paste0(dir, 'SW_AllEncMap.png'), width=8, height = 6, units='in', dpi=300)

```

## October 2020
### Plot all encounters
Shape by enc type (loc, unloc, vis), color by year
```{r}
#Add on a 'year' column
SwEnvData_mapbYr1 <- filter(SwEnvData_mapb, survey == '1303')
SwEnvData_mapbYr2 <- filter(SwEnvData_mapb, survey == '1604')
SwEnvData_mapbYr3 <- filter(SwEnvData_mapb, survey == '1641')
SwEnvData_mapbYr4 <- filter(SwEnvData_mapb, survey == '1705' | survey == '1706')

SwEnvData_mapbYr1$year <- '2013'
SwEnvData_mapbYr2$year <- '2016'
SwEnvData_mapbYr3$year <- '2010'
SwEnvData_mapbYr4$year <- '2017'

SwEnvData_mapb <- rbind.data.frame(SwEnvData_mapbYr1,SwEnvData_mapbYr2,SwEnvData_mapbYr3,SwEnvData_mapbYr4)

SwEnvData_p <- filter(SwEnvData_mapb, pa > 0)
SwEnvData_a <- filter(SwEnvData_mapb, pa == 0)
SwEnvData_a$enctype = 'NA'
SwEnvData_mapb1 <- filter(SwEnvData_p, loc == 1)
SwEnvData_mapb2 <- filter(SwEnvData_p, loc == 0, sid == 999)
SwEnvData_mapb3 <- filter(SwEnvData_p, loc == 0, sid < 999)
SwEnvData_mapb1$enctype <- 'loc'
SwEnvData_mapb2$enctype <- 'unloc'
SwEnvData_mapb3$enctype <- 'both'

SwEnvData_p = as.data.frame(rbind(SwEnvData_mapb1, SwEnvData_mapb2, SwEnvData_mapb3))
SwEnvData_p2 = as.data.frame(rbind(SwEnvData_mapb1, SwEnvData_mapb2))
SwEnvData_mapb = rbind(SwEnvData_p, SwEnvData_a)

hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(SwEnvData_mapb, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 20, col='blueviolet', lwd=2, cex=0.75)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
```

#Load GPS data for tracklines
```{r}
surveynum = 1706
gridEffort <- readRDS(here::here( paste0('output/gridEffort/', gridsize, ' km-', loctype, '/gridEffort', surveynum, '_', gridsize, 'km_', loctype2, '_raw.rda') ))

#Nov 29
gps1641 <- gridEffort$gps
gps1641$year <- 2010
gps1303 <- gridEffort$gps
gps1303$year <- 2013
gps1604 <- gridEffort$gps
gps1604$year <- 2016
gps1705 <- gridEffort$gps
gps1705$year <- 2017
gps1706 <- gridEffort$gps
gps1706$year <- 2017
whalePts <- subset(SwEnvData_mapb, pa >0)
```

```{r}
#"#d64161", "#ff7b25",  "#86af49", "#6b5b95"
ggplot(gps1641, aes(Longitude, Latitude)) +
    geom_point(color='gray66', size=0.5, pch=16) +
  geom_point(gps1303, mapping=aes(Longitude, Latitude),color='gray66',size=0.5, pch=16) +
  geom_point(gps1604, mapping=aes(Longitude, Latitude),color='gray66',size=0.5, pch=16) +
  geom_point(gps1705, mapping=aes(Longitude, Latitude),color='gray66',size=0.5, pch=16) +
  geom_point(gps1706, mapping=aes(Longitude, Latitude),color='gray66',size=0.5, pch=16) +   theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, color=factor(year)), size=1.5) + 
  scale_color_manual(values = c("#0040ff", "#ff7b25",  "#00e6e6", "#ff1ac6"), name='Year') + 
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2)) +
  scale_x_continuous(name='Longitude') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = '#86af49', color = 'black') +
  coord_fixed(1) +
  ggtitle("All Sperm Whale Encounters") + 
      scalebar(x.min = 180, x.max = 185,
                   y.min = 16, y.max = 20, dist = 250, dist_unit = "km",
             transform = TRUE, model = "WGS84")
  
ggsave(paste0(dir, '/SW_TracklineMap.png'), width=8, height = 6, units='in', dpi=300)

```


#Map of all sperm whales
```{r}
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=as.factor(year)), size=1, pch=16) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, shape=factor(enctype)))  +
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2)) +
  scale_x_continuous(name='Longitude') +  #, breaks=round(Longitude, 1) ) +
   scale_shape_manual(values=c(8, 17, 15, 1),name='Enc Type', labels=c("Ac_Localized", "Ac_Trackline","Sighting")) +
 scale_color_manual(values = c("#d64161", "#ff7b25",  "#86af49", "#6b5b95"), name='Year') + #feb236", "#92a8d1",
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("All Sperm Whale Encounters") +
  theme(legend.position="bottom", legend.box='horizontal',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))


ggsave(paste0(dir, 'SW_AllEncMap.png'), width=8, height = 6, units='in', dpi=300)

```
#Plot temp at 600m - October 2020
For email to Phoebe, Acoustic Only encounters included but can be changed using SwEnvData_p
```{r}
long = seq(min(SwEnvData_mapb$Longitude), max(SwEnvData_mapb$Longitude), 5)


ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=temp600m.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p2, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkcyan')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#000099", "#8000ff", "#ff9900"), name=expression(paste("Temp584m ("^"o", "C",")")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Temps at 584 m by Year") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SwMap_T584.png'), width=8, height = 6, units='in', dpi=300)
```
Inspecting the area where the foragers are in the NWHI
Looking at distance to seamounts
```{r}
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=temp600m.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p2, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkcyan')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#000099", "#8000ff", "#ff9900"), name=expression(paste("Temp584m ("^"o", "C",")")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Temps at 584 m by Year") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
```

## For Jeff
Oct 12, 2020- depths of constant temp, don't have, but have depths? - NOT WHAT HE WANTED. WANTED DEPTH OF WATER COLUMN WHERE IT WAS 6DEG C
Acoustics only encounters to be consistent
```{r}

## SSH ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=ssh.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), 
                       name=expression(paste("SSH (m)")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("SSH for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_SSH.png'), width=8, height = 6, units='in', dpi=300)



### SSHsd ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=sshsd.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), name=expression(paste("SSHsd (m)")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("SSHsd for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_SSHsd.png'), width=8, height = 6, units='in', dpi=300)


### EKE ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=eke.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), 
                       name=expression("EKE (m"^2* "/s"^2*")"))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("EKE for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_EKE.png'), width=8, height = 6, units='in', dpi=300)



### CHLa ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=chla.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), 
                       name=expression("Chl-a (mg/m"^3*")"))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Chl-a for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_chla.png'), width=8, height = 6, units='in', dpi=300)


### SST ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=sst.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#0033cc","#66ff66","#ff0000"), 
                       name=expression(paste("SST ("^"o", "C",")")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("SST for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_SST.png'), width=8, height = 6, units='in', dpi=300)




### Wave Power ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=wavepower.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 4),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), 
                       name='WavePower kW/m')  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Wave Power for All Surveys") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))
ggsave(paste0(dir, 'SwMap_WP.png'), width=8, height = 6, units='in', dpi=300)


## Temp at 584m ####
ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=temp600m.r, shape=factor(year)), size=1.2) +
  geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), name=expression(paste("Temp584m ("^"o", "C",")")))  +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Temps at 584 m by Year") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SwMap_T584green.png'), width=8, height = 6, units='in', dpi=300)

#Bathymetry ####

ggplot(SwEnvData_mapb, aes(Longitude, Latitude)) +
  geom_point(aes(color=depth.r, shape=factor(year)), size=1.2) +
  geom_point(PmNonForPres, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +

  theme_bw() +
  # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')   +
 scale_color_gradientn(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"), name="Depth (m)" ) +
  scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Depth") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

## Encounter Rate Foraging ####
PmFor <- readRDS(here::here( paste0('data/Pm_ForagingOnly.rda') ))
PmForAbs <- subset(PmFor, pa == 0)
PmForPres <- subset(PmFor, pa > 0)
ggplot(PmFor, aes(Longitude, Latitude)) +
  geom_point(PmForAbs, mapping = aes(color=pa, shape=factor(year)), size=1.2) +
  geom_point(PmForPres, mapping = aes(color=pa, shape=factor(year)), size=1.2) +
  # geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +
  # scale_color_discrete(name = "Encounter Rate") +  #if color is factor
  scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')  +
  theme_bw()  +
    # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(19, 3, 8, 18),
                     # name='Survey Year')   +
  # scale_fill_manual(values  =c("#AFEEEE", '#00FA9A', '#FF8C00', '#8B0000'), name="Encounter Rate") +
 scale_color_gradientn(colors =c("#AFEEEE", '#FF8C00', '#FF00FF', '#8B0000'), name=expression(paste("Number of Groups per cell")))  +
  # scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Foraging Encounter Rate by Year") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SwMap_EncRate3.png'), width=8, height = 6, units='in', dpi=300)


## Encounter Rate Non-Foraging ####
PmNonFor <- readRDS(here::here( paste0('data/Pm_NonForagingOnly.rda') ))
PmNonForAbs <- subset(PmNonFor, pa == 0)
PmNonForPres <- subset(PmNonFor, pa > 0)
nrow(PmNonForAbs)
ggplot(PmNonFor, aes(Longitude, Latitude)) +
  geom_point(PmNonForAbs, mapping = aes(color=pa, shape=factor(year)), size=1.2) +
  geom_point(PmNonForPres, mapping = aes(color=pa, shape=factor(year)), size=1.2) +
  # geom_point(SwEnvData_p, mapping=aes(Longitude, Latitude, fill='Whales'), pch=10, color='darkmagenta')  +
  # scale_color_discrete(name = "Encounter Rate") +  #if color is factor
  scale_shape_manual(values=c(19, 3, 8, 18),
                     name='Survey Year')  +
  theme_bw()  +
    # scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  # scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(19, 3, 8, 18),
                     # name='Survey Year')   +
  # scale_fill_manual(values  =c("#AFEEEE", '#00FA9A', '#FF8C00', '#8B0000'), name="Encounter Rate") +
 scale_color_gradientn(colors =c("#AFEEEE", '#FF8C00', '#FF00FF', '#8B0000'), name=expression(paste("Number of Groups per cell")))  +
  # scale_fill_manual(values = 10, name='') +
  geom_polygon(data = hawaiiMap, aes(x = Longitude2, y = Latitude, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Non-Foraging Encounter Rate by Year") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 10),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SwMap_NonForEncRate.png'), width=8, height = 6, units='in', dpi=300)





```


This includes different types of encounters
```{r}
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=as.factor(year), shape=enctype), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2)) +
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(8, 17, 15),name='Enc Type', labels=c("Ac_Localized", "Ac_Unlocalized","Sighting")) +
 scale_color_manual(values = c("#d64161", "#ff7b25",  "#86af49", "#6b5b95"), name='Year') + #feb236", "#92a8d1",
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("All Sperm Whale Encounters") +
  theme(legend.position="bottom", legend.box='horizontal',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))


ggsave(paste0(dir, 'SW_AllEncMap.png'), width=8, height = 6, units='in', dpi=300)

```



##Plot temp at 0m
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#for temperature at surface
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=sstAQ_m), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2)) +
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(8, 18),name='Localized?') +
                      # values=c(8, 17, 15, 18, 12, 19),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#000099", "#8000ff", "#ff9900"), name=expression(paste("SST ("^"o", "C",")"))) +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Sea Surface Temperature for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_SSTMap.png'), width=8, height = 6, units='in', dpi=300)
```

##Plot temp at 100m
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#for temperature at surface
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(shape=as.factor(click_code2), color=temp105C), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
                     name='click types ',
                     labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#00AFBB", "#E7B800", "#FC4E07"), name='Temp100 m, degC') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Temps at 100 m") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(here::here('output', 'SW_T100Map.png'), width=8, height = 6, units='in', dpi=300)
```

#Plot temp at 600m
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#for temperature at surface
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=temp600C), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#000099", "#8000ff", "#ff9900"), name=expression(paste("T600m ("^"o", "C",")")))+
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Temps at 600 m for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_T600Map.png'), width=8, height = 6, units='in', dpi=300)
```

#bathymetry
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#for bathy
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=bath), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#000066", "#4d88ff", "#00e600","#ffff00", "#ff9900", "#ff0000"), name='Bathy, m') + #ff9b00 
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Bathymetry for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Bathy.png'), width=8, height = 6, units='in', dpi=300)
```

#SSH plot
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#sea surface height
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=ssh), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#8ad334", "#ff0000"), name='SSH (m)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("SSH for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_SSH.png'), width=8, height = 6, units='in', dpi=300)
```

#SSHsd plot
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#sea surface height
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=sshsd), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#8ad334", "#ff0000"), name='SSH_SD (m)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Standard Deviation of SSH for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_SSHsd.png'), width=8, height = 6, units='in', dpi=300)



#####Compare sshsd#######
#June 6, 2020: Compare matched locations to true locations -> these match up super well

ggplot() +
  geom_point(SwEnvData, mapping =aes(lon2, lat, color=sshsd), size=3) +
  geom_point(data=sshsd_match2, mapping = aes(lon, lat), size=3, pch=1, color='magenta') +
  ggtitle("Comparing Matched to True Locations of SSHSD") +
  theme_bw() 

na<-which(is.na(sshsd_match2))
sshsd_match2[121,]


```

#EKE plot
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#EKE
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=eke), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#1ab2ff", "#008000"), name=expression(paste("EKE (m"^"2", "/s"^"2",")"))) + 
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Eddy Kinetic Energy for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_EKE.png'), width=8, height = 6, units='in', dpi=300)
```


#ChlA plot
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#chla
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=chla_m), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#00AFBB", "#E7B800", "#FC4E07"), name=expression(paste("Chl a (mg", "/m"^"3",")"))) + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Chlorophyll A for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Chla.png'), width=8, height = 6, units='in', dpi=300)
```


#wave power
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


#chla
ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=wavepow), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   # scale_shape_manual(values=c(21, 17, 15, 19, 10, 3),
   #                   name='click types ',
   #                   labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow")) +
 scale_color_gradientn(colors = c("#1ab2ff", "#008000"), name='WP (kW/m)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Wave Power for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_WavePow.png'), width=8, height = 6, units='in', dpi=300)
```

#Aspect
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=asp_deg), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
  scale_color_gradientn(colors = c("#000080","#1a1aff", "#00cc99"), name='Aspect (deg)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Aspect for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Aspect.png'), width=8, height = 6, units='in', dpi=300)
```

#Slope
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=slp_deg), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
  scale_color_gradientn(colors = c("#000080","#1a1aff", "#00cc99"), name='Slope (deg)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Slope for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Slope.png'), width=8, height = 6, units='in', dpi=300)
```

#Distance to Seamounts
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=d2smt_km), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
  scale_color_gradientn(colors = c("#000080","#0000ff", "#bfff00"), name='Dist to Seamt (km)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Distance to Seamounts for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Dist2Smt.png'), width=8, height = 6, units='in', dpi=300)
```

#Distance to Land
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)


ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(color=d2land_km), size=3) +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
  scale_color_gradientn(colors = c("#000080","#0000ff", "#bfff00"), name='Dist to Land (km)') + #c("#fcff00", "#005c8b", "#b04ce5")
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Distance to Land for All Sperm Whales") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_Dist2Land.png'), width=8, height = 6, units='in', dpi=300)
```

#click code plots
Separate these into separate plots per click type
```{r}
long = seq(min(SwEnvData_mapb$lon), max(SwEnvData_mapb$lon), 5)

ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(shape =click_code2, color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(21, 17, 15, 19, 10, 3, 20),
                     name='click types ',
                     labels=c("coda/feed", "coda/feed/slow","coda/slow","feed", "feed/slow", "slow", "coda")) +
 # scale_fill_brewer(palette = "Set1", name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Click Types for All Sperm Whales") +
  theme(legend.position="bottom", legend.box='horizontal',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_EncType.png'), width=8, height = 6, units='in', dpi=300)
```

Use 4 combined click groups
```{r}

#all groups

ggplot(SwEnvData_mapb, aes(lon, lat)) +
  geom_point(aes(shape =as.factor(click_grp), color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
  scale_shape_manual(values=c(19,15,17,18),
                     name='group type',
                     labels=c("forage", "male/forage", "male", "social")) +
  scale_color_manual(values = c("#8cff66", "#005c8b", "#b04ce5"), name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("All Group Types") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_allgrps.png'), width=8, height = 6, units='in', dpi=300)


#foraging only
forage = filter(SwEnvData_mapb, click_grp == 'forage')

ggplot(forage, aes(lon, lat)) +
  geom_point(aes(shape =click_grp, color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(19),
                     name='group type',
                     labels=c("forage")) + #c("forage/male", "male","forage","social"))
  scale_color_manual(values = c("#8cff66", "#005c8b", "#b04ce5"), name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Foraging Groups (creaks and regular clicks)") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_forage.png'), width=8, height = 6, units='in', dpi=300)

#males
male = filter(SwEnvData_mapb, click_grp == 'male')

ggplot(male, aes(lon, lat)) +
  geom_point(aes(shape =click_grp, color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(17),
                     name='group type',
                     labels=c("male")) + #c("forage/male", "male","forage","social"))
  scale_color_manual(values = c("#8cff66", "#005c8b", "#b04ce5"), name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Solo Males (slow clicks only)") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_male.png'), width=8, height = 6, units='in', dpi=300)


#forage/males
maleforg = filter(SwEnvData_mapb, click_grp == 'forage/male')

ggplot(maleforg, aes(lon, lat)) +
  geom_point(aes(shape =click_grp, color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(15),
                     name='group type',
                     labels=c("male/forage")) + #c("forage/male", "male","forage","social"))
  scale_color_manual(values = c("#8cff66", "#005c8b", "#b04ce5"), name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Solo Males & Foragers (regular and slow clicks)") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_maleforg.png'), width=8, height = 6, units='in', dpi=300)


#social groups
social = filter(SwEnvData_mapb, click_grp == 'social')

ggplot(social, aes(lon, lat)) +
  geom_point(aes(shape =click_grp, color =enctype), size=3)+#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Latitude',breaks=seq(0,30,2))+
  scale_x_continuous(name='Longitude', breaks=round(long,1)) +
   scale_shape_manual(values=c(18),
                     name='group type',
                     labels=c("social")) + #c("forage/male", "male","forage","social"))
  scale_color_manual(values = c("#8cff66", "#005c8b", "#b04ce5"), name='Enc Type') +
  geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  coord_fixed(1) +
  ggtitle("Social Groups (includes codas)") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE))

ggsave(paste0(dir, 'SW_ClickType_social.png'), width=8, height = 6, units='in', dpi=300)
```


## Clicks vs Group size
```{r}
# plot(SwEnvData_mapb3$cpm, SwEnvData_mapb3$grpsize, xlab='Clicks per Minute', ylab='Group Size')

SwEnvData_mapb4 <- filter(SwEnvData_mapb, cpm != 'Inf')
mdlclick <- lm(SwEnvData_mapb4$grpsize ~ SwEnvData_mapb4$cpm)
pvalmdlclick = round(summary(mdlclick)$coefficients[2,4],5)
r2mdlclick = summary(mdlclick)$adj.r.squared

plotOrder <- order(SwEnvData_mapb4$cpm)
cint <- as.data.frame(predict(mdlclick, interval = "confidence"))

ggplot(SwEnvData_mapb, aes(cpm, grpsize)) +
  stat_summary(fun.data = mean_cl_normal) +
  geom_smooth(method = 'lm') +
  geom_point(aes(shape =click_grp, color =as.factor(survey)), size=3) +#, color = 'blue') +
  theme_bw() +
  scale_y_continuous(name='Group Size')+
  scale_x_continuous(name='Clicks per Minute')+
  scale_shape_manual(values=c(19,15,17,18),
                     name='group type',
                     labels=c("forage", "male/forage", "male", "social")) +
   scale_color_manual(values = c("#d64161", "#e67300", "#feb236", "#86af49","#92a8d1", "#6b5b95"), name='Survey')+
  # geom_polygon(data = hawaiiMap, aes(x = long, y = lat, group = group), fill = 'black', color = 'black') +
  # coord_fixed(1) +
  # ggtitle("# Clicks per min vs Group Size") +
  theme(legend.position="right", legend.box='vertical',legend.text = element_text(size = 8),legend.title = element_text(size=9),plot.title = element_text(hjust = 0.5))+
  guides(shape = guide_legend(nrow = 2, byrow = TRUE)) 
  




```

#Correlation Matrix
```{r}
##Calculate Pearson's correlation coefficient for all variables. Remove variables -0.70 > r > 0.70.
#use only the env predictos
require(corrplot)
require(PerformanceAnalytics) 
require(RColorBrewer)

sw_scaled <- SwEnvData_scale[,27:42]
sw_orig <- SwEnvData[,27:41]

corrplot.mixed(cor(sw_scaled, use='complete.obs'), upper="circle", lower="number", upper.col =brewer.pal(n=8, name="PRGn"), lower.col = brewer.pal(n=8, name="PRGn") )
tmp <- cor(sw_scaled, use='complete.obs')
tmp[upper.tri(tmp)] <- 0
diag(tmp) <- 0
sw_new <- sw_scaled[,!apply(tmp,2,function(x) any(abs(x) > 0.70))]

#two ways to make corr plot
corrplot(cor(sw_new, use="complete.obs"),type = 'lower', col=brewer.pal(n=8, name="PRGn")) 

corrplot.mixed(cor(sw_new, use='complete.obs'), upper="circle", lower="number", upper.col =brewer.pal(n=8, name="PRGn"), lower.col = brewer.pal(n=8, name="PRGn") )

# chart.Correlation(sw_new, histogram=TRUE, pch=19)

sw_cor <- cbind(dplyr::select(SwEnvData, ID:lon_trk2), sw_new)

write.csv(sw_cor, here::here('output', 'SpermiesUncorrelated_20200527.csv'), row.names = F)



```
