==> code/9_DataCleanup.Rmd <==
---
title: "Environmental Data Inspection"
author: "Yvonne Barkley"
date: "7/27/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r}
library(tidyverse)

#to remove variables, replace **
# rm(list = ls()[grep("wil", ls())])
```



# Check environmental data

#load data for a survey
```{r}
surveynum = 1604
gridsize = 10
loctype = 'localized'
loctype2 = 'loc'


# Get gridEffort for the correct survey and spatial scale. 
gridEffortCleanup <- readRDS(paste0(here::here('output'), '/', 'gridEffort', surveynum, '_', gridsize, 'km.rda'))
pmDetsGrid <- gridEffortCleanup$detections
pmDetsGrid <- pmDetsGrid[, c(1:26,43,45)]
```

```{r}
env1_sst      <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_sst.rda')))
env2_chla     <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_chla.rda')))
env3_godas600 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_temp600.rda')))
env4_ww3      <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_wavepow.rda')))
env5_bath     <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_bathy.rda')))
env67_slpasp  <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_slpasp.rda')))
env8_d2land   <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_d2land.rda')))
env9_ssh      <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_ssh.rda')))
env10_sshsd   <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_sshsd.rda')))
env11_eke     <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_eke.rda')))
env12_d2smt   <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_dist2smt.rda')))
# env12_d2smt   <- read.csv(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0(surveynum, '_', gridsize, 'km_dist2smt.csv')))
```

A little extra clean-up
```{r}
env5_bath = filter(env5_bath, bath_m < -100)
env12_d2smt   <- filter(env12_d2smt, effCells %in% env5_bath$effCells)
env1_sst   <- filter(env1_sst, effCells %in% env5_bath$effCells)
env2_chla   <- filter(env2_chla, effCells %in% env5_bath$effCells)
env3_godas600 <- filter(env3_godas600, effCells %in% env5_bath$effCells)
env4_ww3 <- filter(env4_ww3, effCells %in% env5_bath$effCells)
env67_slpasp <- filter(env67_slpasp, effCells %in% env5_bath$effCells)
env8_d2land <- filter(env8_d2land, effCells %in% env5_bath$effCells)
env9_ssh <- filter(env9_ssh, effCells %in% env5_bath$effCells)
env10_sshsd <- filter(env10_sshsd, effCells %in% env5_bath$effCells)
env11_eke <- filter(env11_eke, effCells %in% env5_bath$effCells)


saveRDS(env12_d2smt, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_dist2smt.rda')))
saveRDS(env1_sst, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_sst.rda')))
saveRDS(env2_chla, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_chla.rda')))
saveRDS(env3_godas600, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_temp600.rda')))
saveRDS(env4_ww3, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_wavepow.rda')))
saveRDS(env67_slpasp, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_slpasp.rda')))
saveRDS(env8_d2land, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_d2land.rda')))
saveRDS(env9_ssh, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_ssh.rda')))
saveRDS(env10_sshsd, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_sshsd.rda')))
saveRDS(env11_eke, here::here('output/data_output/10 km-localized', paste0(surveynum, '_', gridsize, 'km_eke.rda')))

```

Change the column names to UTC
```{r}
colnames(env3_godas600)[colnames(env3_godas600) == "date"] <- "UTC"
colnames(env4_ww3)[colnames(env4_ww3) == "date"] <- "UTC"
colnames(env9_ssh)[colnames(env9_ssh) == "date"] <- "UTC"
colnames(env10_sshsd)[colnames(env10_sshsd) == "date"] <- "UTC"
colnames(env11_eke)[colnames(env11_eke) == "date"] <- "UTC"

```


Fill NAs to see where they are
```{r}
matchedData[matchedData[,8]=='NA'] <- NA                  #put in NaNs to variable column

env1_sst[env1_sst[,8]=='NA'] <- NA      
env2_chla[env2_chla[,8]=='NA'] <- NA       
env3_godas600[env3_godas600[,5]=='NA'] <- NA  
env4_ww3[env4_ww3[,9]=='NA'] <- NA        
env5_bath[env5_bath[,1]=='NA'] <- NA
env67_slpasp[env67_slpasp[,3]=='NA'] <- NA    
env67_slpasp[env67_slpasp[,4]=='NA'] <- NA    
env8_d2land[env8_d2land[,1]=='NA'] <- NA     
env9_ssh$ssh[env9_ssh[,4]=='NA'] <- NA     
env10_sshsd$sshsd[env10_sshsd[,4]=='NA'] <- NA     
env11_eke$eke[env11_eke[,4]=='NA'] <- NA       
env12_d2smt[env12_d2smt[,2]=='NA'] <- NA 


env11_eke$eke[is.na(env11_eke[, 4]),]


```


#Bind data
Bind all data together with the presence/absence information
```{r}

PmEnvTMP <- data.frame(env1_sst[, c(1:3, 5:8)], `chla`=env2_chla[, 8], `temp600`=env3_godas600[,5], `wavepow`=env4_ww3[,9], `bathy`=env5_bath[,1], `slope`=env67_slpasp[,3], `aspect`=env67_slpasp[,4], `d2land`=env8_d2land[,1], `ssh`=env9_ssh[,4], `sshsd`=env10_sshsd[,4], `eke`=env11_eke[,4], `d2smt`=env12_d2smt[,2])
colnames(PmEnvTMP)[colnames(PmEnvTMP) == "sst_mean"] <- "sst"

#primarily for the 10km loc datasets 1604 
# PmEnvTMP <- filter(PmEnvTMP, PmEnvTMP$bath_m < -100 )


```

#Remove NAs
Check for NAs with pa=1 and pa=0
```{r}

#find any NAs separately in presences or absences before removing them
pa1 <- filter(PmEnvTMP, pa == 1)
pa0 <- filter(PmEnvTMP, pa == 0)
pa0$survey <- surveynum

# pa1[is.na(pa1),]
# pa0[is.na(pa0),]
sum(is.na(pa1))     # Total NA for presences
colSums(is.na(pa1)) # Total NA per column
sum(is.na(pa0))     # Total NA for absences
colSums(is.na(pa0)) # Total NA per column


PmEnv <- PmEnvTMP %>% drop_na()
PmEnvAbs <- filter(PmEnv, pa == 0)
PmEnvPres <- filter(PmEnv, pa == 1)
# PmEnvPres <- rename(PmEnvPres, Longitude = lon, Latitude=lat)
# PmEnvAbs <- rename(PmEnvAbs, Longitude = lon, Latitude=lat)

```


#Merge data
Get detection numbers and click info for presences. Everything was saved in the '$detections' part of the list for gridEffort. 
Match up the click info with PmEnv using the gridIndex and effCells (named effCells in 8_GetEnvData.Rmd)
```{r}
# connect the centroids for Pm detections and env data with click info from pmDetsGrid
merge_tmp <- merge(PmEnvPres, pmDetsGrid, by.x = 'effCells', by.y = 'gridIndex')
merge_tmp2 <- dplyr::select(merge_tmp, c(2,23,19:22,1,3,45,4,5,28,27,everything())) #This contains the presences with the env data from the centroid locations
# spruce this up some more to bind better with the absences

merge_tmp2 <- merge_tmp2 %>% rename(UTC=UTC.x, UTCp = UTC.y, Longitude=Longitude.x, Latitude = Latitude.x, Latitudep = Latitude.y, Longitudep = Longitude.y)

trythis <- bind_rows(merge_tmp2, PmEnvAbs)
trythis$survey = surveynum

(dataname <- paste0(surveynum, '_', gridsize, 'km_', loctype2, '_complete'))
assign(dataname, trythis)

#save completed file
saveRDS(assign(dataname, trythis), here::here(paste0('output/output/', gridsize, ' km-', loctype), dataname))



```


#Load and bind data from all surveys
```{r}
Complete1641 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('1641_', gridsize, 'km_', loctype2, '_complete')))

Complete1303 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('1303_', gridsize, 'km_', loctype2, '_complete')))

Complete1604 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('1604_', gridsize, 'km_', loctype2, '_complete')))

Complete1705 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('1705_', gridsize, 'km_', loctype2, '_complete')))

Complete1706 <- readRDS(here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('1706_', gridsize, 'km_', loctype2, '_complete')))


CompleteTotal <- bind_rows(Complete1641, Complete1303, Complete1604, Complete1705, Complete1706)

sum(is.na(CompleteTotal[,15:26]))     # Total NA for env data columns
colSums(is.na(CompleteTotal[,15:26])) # Total NA per column


saveRDS(CompleteTotal, here::here(paste0('output/envData/', gridsize, ' km-', loctype), paste0('CompletePm_', gridsize, 'km_', loctype2, '_raw.rda')))

```



#Scale data
```{r}
# SCALE and CENTER PREDICTORS

CompletePreds <- dplyr::select(CompleteTotal, sst:d2smt)
CompleteScaledtmp <- scale(CompletePreds, scale = TRUE)
CompleteScaled <- data.frame(dplyr::select(CompleteTotal, UTC:pa), CompleteScaledtmp, dplyr::select(CompleteTotal, peak:lon_trk2))
write.csv(CompleteScaled, here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('CompletePm_', gridsize, 'km_', loctype2, '_scaled.csv')), row.names = FALSE)

saveRDS(CompleteScaled, here::here(paste0('output/data_output/', gridsize, ' km-', loctype), paste0('CompletePm_', gridsize, 'km_', loctype2, '_scaled.rda')))


```
