==> code/10_GAMS_Forager_25km.Rmd <==
---
title: "GAMs for Foraging Whales"
author: "Yvonne Barkley"
date: "10/4/2020"
output:
  pdf_document: default
  html_document: default
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment = '', fig.width =8, fig.height = 6, message=FALSE, tidy.opts=list(width.cutoff=60),tidy=TRUE)
```

Load libraries
```{r, echo=T, eval=FALSE}
library(tidyverse)
library(mgcv)
library(corrplot)
library(geoR)
library(here)
```


# Research question:
## What environmental variables characterize sperm whale habitat? 
## Hypothesis: Sperm whales are found in deep, productive offshore waters. 



Load universal variables
```{r}
#Values used for file and directory names
survey = 'AllSurveys'
gridsize = 25
loctype = 'Combined'
loctype2 = 'Comb'

```

Load data from 'models/data' folder
```{r}
PmScaled <- readRDS(here::here( paste0('output/models/', loctype, '/data/', 'CompletePm_', gridsize, 'km_', loctype2, '_scaled.rda') ))
# add column for log effort as offset #
PmScaled$log.effort	= log(PmScaled$EffArea)
PmScaled <- subset(PmScaled, chla <= 9) #some outliers in a handful of absences
PmScaled$distseamt.r=  PmScaled$distseamt.r/1000 
```
```{r eval=FALSE,echo=FALSE}

#1. Denote forage or not. 
PmFor1 <- filter(PmScaled, grepl('r|k', click_code))   # the 'k' implies ck for creaks, c also for codas
PmFor0 <- filter(PmScaled, !grepl('r|k', click_code))
PmFor1$forage <- 1  #forage
PmFor0$forage <- 0  #non-forage
PmForage <- bind_rows(PmFor1, PmFor0)

nrow(PmFor0)
nrow(PmFor1)

#2. Denote social group or not

PmGroup1 <- filter(PmScaled, grepl('cd', click_code))
PmGroup0 <- filter(PmScaled, !grepl('cd', click_code))

#designate whether the encounters included codas (1) or not (0)
PmGroup1$group <- 1
PmGroup0$group <- 0
PmGroup <- rbind(PmGroup1, PmGroup0)

```

Check correlation of covariates
```{r fig.width =10, fig.height = 6}
require(corrplot)
corrplot.mixed(cor(PmScaled[,18:29]), upper="number", lower="circle")

# Are all correlation coefficients < |0.6|?
abs(cor(PmScaled[,18:29])) <= 0.6

```

### KS tests 
I compared the distributions of environmental data between the whales and the absences. 
Plots are attached in separate powerpoint. In summary, temperature at 600 m, SSH, and chlorophyll were the only variables with significantly different distributions (p-value < 0.05). However, the D statistics were close to zero (D ~ 0.1) for each, indicating that although the distributions were different, they were not that far apart. The plots also show how similar the general shape of the distributions are between where the whales were observed and where they were absent.
```{r echo=FALSE, eval=FALSE}
# par(mfrow = c(1,1))
# with(CompleteTotal, plot(pa~bath_m))
require(Cairo)
require(ggplot2)
source('./code/multiplot.r') 

 
loopVec <- 18:29  #columns from PmScaled to plot
 
par(mfrow = c(1, 3), mar=c(3,3,1,1), oma=c(0,0,3,1))
 
for (j in loopVec){
   #use original values for plots with PmRaw
   datPlot <- PmScaled[,c(1,j)]  #get dataframe of pa and variable j
   datPlot$bin <- 0
   datPlot$bin <- ifelse(datPlot$pa >0 , 1, 0) #put 1s in color col for plotting later
   datPlot <- select(datPlot, bin, 2, pa)
    #normal plot
   # plot(datPlot[,2], datPlot[,1], xlab = paste('scaled',colnames(datPlot)[2]), ylab = 'whales per cell')
   

   ptplt <-  ggplot(datPlot, aes(datPlot[,2],datPlot[,1], colour=as.factor(datPlot[,1]))) +
       geom_point() +   
     theme(legend.position ="none") +
       xlab(paste(colnames(datPlot)[2], 'scaled' )) +
       ylab('whales per cell')+
       # scale_colour_discrete(name='',labels = c('Whales Absent', 'Whales Present'))+
theme_bw() +
  guides(colour=FALSE)
   
   #histogram
   # hist(datPlot[,2], xlab = paste('scaled',colnames(datPlot)[2]), ylab = 'frequency', main='' )
   
   histplt <- ggplot(datPlot, aes(datPlot[,2], fill=as.factor(datPlot[,1]))) +
     geom_histogram(colour='black', binwidth = 0.5) +
    xlab(paste(colnames(datPlot)[2],'scaled' )) +
    ylab('Frequency') +
     scale_fill_discrete(name = '', labels = c('Whales Absent', 'Whales Present'))+
     theme_bw() +
     theme(legend.position ="top",
           legend.text=element_text(size=16)) 

   
   #do KS test and plot
   datPlot1 <- filter(datPlot, pa >0)
   datPlot0 <- filter(datPlot, pa ==0)

   datKS <- ks.test(datPlot1[,2], datPlot0[, 2])
     cdf0 <- ecdf(datPlot0[,2]); cdf1 <- ecdf(datPlot1[,2]) #make cdfs for the absences and whales for that variable
   #find min and max stats to draw the line between the greatest distance between the distributions
   minMax <- seq(min(datPlot0[,2], datPlot1[,2]), max(datPlot0[,2], datPlot1[,2]), length.out=length(datPlot0[,2])) 
   x0 <- minMax[which( abs(cdf0(minMax) - cdf1(minMax)) == max(abs(cdf0(minMax) - cdf1(minMax))) )] 
   y0 <- cdf0(x0) #sometimes there are 2 points with the same minmax, take the first one 
   y1 <- cdf1(x0)
   
  ksplt <- ggplot(datPlot, aes(x = datPlot[,2], group = datPlot[,1], color = as.factor(datPlot[,1]))) + #
  stat_ecdf(size=1) +
    theme_bw() +
    theme(legend.position ="none", plot.title = element_text(hjust= 0.5)) +
    xlab(paste(colnames(datPlot)[2],'scaled') ) +
    ylab("ECDF") +
    #geom_line(size=1) +
    geom_segment(aes(x = x0[1], y = y0[1], xend = x0[1], yend = y1[1]),
        linetype = "dashed", color = "red") +
    geom_point(aes(x = x0[1] , y= y0[1]), color="red", size=2) +
    geom_point(aes(x = x0[1] , y= y1[1]), color="red", size=2) +
    ggtitle("K-S Test") +
    # theme(legend.title=element_blank()) +
     scale_colour_discrete(name = 'Distribution Type', labels = c('Whales Absent', 'Whales Present'))+
    annotate("text",x=max(datPlot[,2])-2,y=min(datPlot[,1])+0.1, hjust=.2,label=paste('D=',round(datKS[1]$statistic,2))) +
    annotate("text",x=max(datPlot[,2])-2,y=min(datPlot[,1])+0.05, hjust=.2,label=paste('p=',round(datKS$p.value,3)))

  
Cairo(here::here('figures', paste0('DataPlotsScaled_', gridsize, 'km_', loctype,'_', colnames(datPlot)[2], '.png')), width=14,height=6,
    units='in', dpi=300, bg='white')
multiplot(ptplt, histplt, ksplt, cols=3)
dev.off()
  
   # ggsave(paste('DataPlots_',colnames(datPlot)[2],'.png'))
  
  }

```


### Data Visualization
Histograms showing the general distribution of each environmental predictor for the entire dataset.
```{r}
par(mfrow = c(3, 4), mar=c(3,3,2,1), oma=c(0,0,3,1))
 
dataSet = PmScaled   #raw values

loopVec <-  30:41  #columns from PmScaled to plot
 
 for (j in loopVec){
   
   datPlot <- dataSet[, c(1,j)]
   
   hist(datPlot[,2], main = colnames(datPlot)[2], ylab='frequency', xlab = '')
      # plot(datPlot[,2], datPlot[,1], ylab = 'Whales', xlab = colnames(datPlot)[2])
mtext(paste0("Acoustics Only Data, ", gridsize, 'km grid'), side=3, line=1, outer=TRUE, cex=1, font=1)

 }

# dev.off()
```



### Data Splitting 
Split the data into train and test sets
```{r eval=FALSE}
require(dplyr)
splitdf <- function(dataframe, seed=NULL) {
    if (!is.null(seed)) set.seed(seed)
    index <- 1:nrow(dataframe)
    trainindex <- sample(index, trunc(length(index)*0.7))
    trainset <- dataframe[trainindex, ]
    testset <- dataframe[-trainindex, ]
    list(trainset=trainset,testset=testset)
}

trainForg = NULL
testForg = NULL
# preschk = NULL
seed = 3

for (s in c(1641, 1303, 1604, 1705, 1706)){
   
 trSub <- filter(PmForage, survey == s)

 #subset for presences and split 70/30
 pres1 <- filter(trSub, pa > 0 & forage == 1)# & loc == 1) #include all presence data/acoustic encounters
 listPres  <- splitdf(pres1, seed) #output is list for train and test
# preschk = rbind(preschk, pres1)
 #subset for absences and split 70/30
 abs0  <- filter(trSub, pa == 0 )
 listAbs <- splitdf(abs0, seed)  #output is list for train and test

 #combine train data for presence and absence
 trainAll <- rbind( listPres$trainset, listAbs$trainset )
 
 #combine test data for presence and absence
 testAll  <- rbind( listPres$testset,  listAbs$testset  )
   
trainForg = rbind( trainForg, trainAll )
testForg  = rbind( testForg,  testAll )

# trainAcOnly$log.effort <- log(trainAcOnly$EffArea)
# testAcOnly$log.effort <- log(testAcOnly$EffArea)
}
saveRDS(trainForg, here::here(  paste0('output/models/',loctype, '/data/Train_', gridsize, 'km_', loctype2, '_Hunt.rda')  ))
saveRDS(testForg, here::here(  paste0('output/models/',loctype, '/data/Test_', gridsize, 'km_', loctype2, '_Hunt.rda')  ))

# nrow(dplyr::filter(trainAcOnly, trainAcOnly$pa >0))
# nrow(dplyr::filter(testAcOnly, testAcOnly$pa >0))
```

## Generalized Additive Models
The data are treated as count data, number of sperm whale encounters per cell, and we used the negative binomial distribution to model the response variable for comparison with the Tweedie distribution. 
We used thin-plate regression splines (the default basis) for the smoothers of the environmental predictors. Each smoother was limited to 3 degrees of freedom (k=3) to reduce overfitting parameters per recommendations from other studies building similar types of cetaceans distribution models.The log of the effort was included as an offset to account for the variation in effort per cell.


### 25 km spatial scale

* NEGATIVE BINOMIAL DISTRIBUTION
* Knots contrained to k=3 according to literature on cetacean distribution models.
* Automatic term selection is uses an additional penalty term when determining the smoothness of the function ('select' argument = TRUE)..
* We excluded all non-significant variables (alpha=0.05) and refit the models until all variables were significant.
* REML is restricted maximum likelihood used to optimize the parameter estimates.

Load training and test data
```{r}
#seed 1
trainHunt <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Train_',   gridsize, 'km_', loctype2, '_Hunt.rda')  ))
testHunt <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Test_',     gridsize, 'km_', loctype2, '_Hunt.rda')  ))

```

# __Model Selection__

## SET 1
### Full Models
    + does not include spatial smoother
    + does not include slope or aspect due to the variation between left and rightes not include slope or aspect due to the variation between left and right


```{r}
require(mgcv)
nbHunt <- gam(pa ~ s(bath_m, k=3) + s(dist, k=3)  + s(d2smt, k=3) +  s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(eke, k=3) + s(wavepow, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(nbHunt)
AIC(nbHunt)    
```
```{r echo=FALSE}
# model diagnostics
par(mfrow=c(2,2))
gam.check(nbHunt)
```

```{r echo=FALSE}
par(mar=c(4,4,3,3),mfrow = c(4,4))
plot(nbHunt, pages = 2, residuals = FALSE, pch = 20, cex = 0.25,
scheme = 1, shade = T, shade.col = 'orange', all.terms = TRUE, main='NegBin Full Foragers')
```


### Trying the Tweedie for comparison
```{r}
twHunt <- gam(pa ~ s(bath_m, k=3) + s(dist, k=3)  + s(d2smt, k=3) +  s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(eke, k=3) + s(wavepow, k=3) + offset(log.effort), data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(twHunt)
AIC(twHunt)        
```

```{r echo=FALSE, eval=FALSE}
par(mar=c(4,4,3,3),mfrow = c(4,4))
plot(twHunt, pages = 2, residuals = FALSE, pch = 20, cex = 0.25,
scheme = 1, shade = T, shade.col = 'hotpink', all.terms = TRUE, main='Tweedie Full Combined-25km')
```

```{r echo=FALSE, eval=FALSE}
# model diagnostics
par(mfrow=c(2,2))
gam.check(twHunt)
```


### Reduced Models
   * Negative Binomial -> higher explained deviance
   * Removed non-significant variables:   
      + depth
      + distance to land
      + distance to seamount
      + SSH
      + eke
      + wave power
```{r}
# * Does NOT include sighted acoustic encounters

nbHuntb <- gam(pa ~ s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(sshsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(nbHuntb)
AIC(nbHuntb)  #slightly lower AIC
```
```{r echo=FALSE}
# model diagnostics
par(mar=c(4,4,3,3),mfrow = c(2,2))
gam.check(nbHuntb)
```

```{r echo=FALSE}
par(mar=c(4,4,3,3),mfrow = c(3,2))
plot(nbHuntb, residuals = FALSE, pch = 20, cex = 0.25, ylim = c(-6,2),
scheme = 1, shade = T, shade.col = 'hotpink', all.terms = TRUE, main='NegBin Reduced Combined-25km')
#### Explained Deviance = `round(((twCombb$null.deviance-twCombb$deviance)/twCombb$null.deviance)*100, 1)`

```



## SET 2
### Full Models: Includes s(Longitude,Latitude)
Includes 2D Lat-Lon smoother to account for spatial structure in the data and fit the spatial variation not explained by the other predictors  


```{r}
nbHuntLL <- gam(pa ~ s(Longitude, Latitude) + s(bath_m, k=3) + s(dist, k=3) + s(d2smt, k=3) +  s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(eke, k=3) + s(wavepow, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(nbHuntLL)
AIC(nbHuntLL)
```
```{r echo=FALSE}
# model diagnostics
par(mar=c(4,4,3,3),mfrow = c(2,2))
gam.check(nbHuntLL)
```

```{r echo=FALSE}
par(mfrow = c(2, 2), mar=c(4,3,2,1))
# par(mar=c(4,4,3,3),mfrow = c(4,4))
plot(nbHuntLL,residuals = FALSE, pch = 20, cex = 0.01,
shade = T, scheme =2, theta= 45, phi=40, shade.col = 'hotpink', all.terms = TRUE)
 
```

```{r echo=FALSE, fig.cap = "Purple dots represent acoustically detected encounters. Black dots are all data points(grid centroids)"}  
par(mfrow = c(1,1))
plot(nbHuntLL, select = 1, scheme = 2, lwd = 2, main='2D Spatial Smoother') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 20, col='blueviolet', lwd=2, cex=0.75)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
```


### Checking Tweedie for comparison
```{r}
twHuntLL <- gam(pa ~ s(Longitude, Latitude) + s(bath_m, k=3) + s(dist, k=3) + s(d2smt, k=3) +  s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(eke, k=3) + s(wavepow, k=3) + offset(log.effort), data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(twHuntLL)
AIC(twHuntLL)
```


### REDUCE MODEL PARAMETERS  
   * Negative Binomial: higher explained deviance, lower AIC than Tweedie
   * Keep:   
      + Lon, Lat
      + SST
      + chlorophyll
      + temp at 600 m
      + SSHsd
```{r echo=FALSE, eval=FALSE}
nbHuntLLb <- gam(pa ~ s(Longitude, Latitude) + s(sst, k=3) + s(chla, k=3) + s(temp600, k=3) + s(sshsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(nbHuntLLb)
```

```{r}
colnames(trainHunt)[30] <- 'SST'
colnames(trainHunt)[31] <- 'Chla'
colnames(trainHunt)[32] <- 'Temp600m'
colnames(trainHunt)[39] <- 'SSHsd'
nbHuntLLb2 <- gam(pa ~ s(Longitude, Latitude) + s(SST, k=3) + s(Chla, k=3) + s(Temp600m, k=3) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(nbHuntLLb2)
```
```{r echo=FALSE}
par(mar=c(4,4,1,3),mfrow = c(2,2))
# plot(nbHuntLLb2, select=1, residuals = FALSE, pch = 20, cex = 0.001,
# shade = T, scheme = 2, theta= 30, phi=20, shade.col = 'orange', all.terms = TRUE)

plot(nbHuntLLb2, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'orange', all.terms = TRUE, xlab='Chl a (mg m-3)' , main='Best Set 2 Models w/Spatial Smoother-Reduced')

plot(nbHuntLLb2, select=3, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'orange', all.terms = TRUE, xlab='Temp at 600 m (°C)', main='Best Set 2 Models w/Spatial Smoother-Reduced')

plot(nbHuntLLb2, select=4, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'orange', all.terms = TRUE, xlab='SSHsd (m)', main='Best Set 2 Models w/Spatial Smoother-Reduced')


```

```{r echo=FALSE, fig.cap = "Purple dots represent acoustically detected encounters. Black dots are all data points." }  
par(mfrow = c(1,1))
plot(nbHuntLLb2, select = 1, scheme = 2, lwd = 2, main='Spatial Smoother') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 20, col='blueviolet', lwd=2, cex=0.75)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
```
$~$



### Conclusions


$~$



## Predict Test Data
```{r tidy=TRUE}

require(magrittr)
require(dplyr)

#### For twCombc, no spatial smoother ####
nbTrainFinal <- trainHunt %>% mutate(resid = resid(nbHuntb), predict = predict(nbHuntb))
predTrain <- predict.gam(nbHuntb, type = 'response', se.fit=TRUE)  #calculate MSE for these to compare with test set. If they're super different, speaks to the genrality of the model
nbTrainFinal$fit <- predTrain$fit
nbTrainFinal$se.fit <- predTrain$se.fit
#using scale of 0,1,2 makes this hard to interpret
nbMSEtrain <- mean((nbTrainFinal$pa - nbTrainFinal$fit)^2)  #MSE


# mean(abs((nbTrainFinal$pa - nbTrainFinal$fit)))  #Mean absolute error
## Calculate MSE AFTER transforming the predictions back to the same scale as the observed data
colnames(testHunt)[30] <- 'SST'
colnames(testHunt)[31] <- 'Chla'
colnames(testHunt)[32] <- 'Temp600m'
colnames(testHunt)[39] <- 'SSHsd'
nbPred <- predict.gam(nbHuntb, newdata = testHunt, type = 'response', se.fit = TRUE)
nbTestFinal <- data.frame(testHunt, fit = nbPred$fit, se.fit=nbPred$se.fit)
nbMSEtest <- mean((nbTestFinal$pa - nbTestFinal$fit)^2) #MSE

# mean(abs((testFinal$pa - testFinal$fit))) #Mean absolute error


#### For nbCombcLL, with spatial smoother ####
#pulling the prediction and residual data from the model
nbTrainLL <- trainHunt %>% mutate(resid = resid(nbHuntLLb2), predict = predict(nbHuntLLb2))
predTrainLL <- predict.gam(nbHuntLLb2, type = 'response')  #calculate MSE for these to compare with test set. If they're super different, speaks to the genrality of the model
nbTrainLL$fit <- predTrainLL

#using scale of 0,1,2 makes this hard to interpret
nbMSEtrainLL <- mean((nbTrainLL$pa - nbTrainLL$fit)^2)  #MSE
# mean(abs((nbTrainFinal$pa - nbTrainFinal$fit)))  #Mean absolute error
## Calculate MSE AFTER transforming the predictions back to the same scale as the observed data

nbPredLL <- predict.gam(nbHuntLLb2, newdata = testHunt, type = 'response', se.fit = TRUE)
nbTestLL <- data.frame(testHunt, fit = nbPredLL$fit, se.fit=nbPredLL$se.fit)
nbMSEtestLL <- mean((nbTestLL$pa - nbTestLL$fit)^2) #MSE

# mean(abs((testFinal$pa - testFinal$fit))) #Mean absolute error

# AIC
nbAIC <- AIC(nbHuntb)
nbAICLL <- AIC(nbHuntLLb2)

# Explained Deviance
nbExpDev = round(((nbHuntb$null.deviance-nbHuntb$deviance)/nbHuntb$null.deviance)*100, 2)
nbExpDevLL = round(((nbHuntLLb2$null.deviance-nbHuntLLb2$deviance)/nbHuntLLb2$null.deviance)*100, 2)

```

```{r}
# make summary table of metrics

table = matrix(NA, nrow = 2, ncol = 5)
colnames(table) = c("Best Models", "ExpDev", "AIC", "MSEtrain", "MSEtest")

# enter info by row

table[1,] <- c("nbCombb", paste0(nbExpDev, '%'), round(nbAIC, 2), round(nbMSEtrain,3), round(nbMSEtest,3))

table[2,] <- c("nbCombLLb2 (w/ s(Lon,Lat))", paste0(nbExpDevLL, '%'), round(nbAICLL,2), round(nbMSEtrainLL,3), round(nbMSEtestLL,3))
require(knitr)
kable(table, caption = "Negative Binomial Model Summary Metrics")

```
